==> source/STEP_04_tables.rmd <==
---
title: "MA optimism - tables"
author: "ML"
date: "7/1/2018"
output: word_document: 
    reference_docx: "si.style.docx"
---

# MA of cognitive bias (optimism) in animal studies.

## STEP4 - making results tables
[do not knit]    


```{r setup, include=F, eval = T}
knitr::opts_chunk$set(echo = F, warning = F, eval = F)
#sessionInfo()
options(scipen=100)

# loading packages
pacman::p_load(metafor, dplyr, readxl, MCMCglmm, knitr, mice)

# load custom functions
source("./functions/functions.R") #functions specific to optimism project (data transformations) and I2 function
```

```{r load data, eval=TRUE, include=FALSE} 
rm(list=ls()) # Clean up
dat <- read.csv("./data/MA_dataset_with_ES_subsets.csv") #for kntting
#dat <- read.csv("data/MA_dataset_with_ES_subsets.csv")
dim(dat) #459 76
str(dat) 
names(dat)

#### SUBSET with only ambiguous cue data points:
dat_amb <- dplyr::filter(dat, ScalePoint %in% c("MID","NN","NP"))
dim(dat_amb) #269

### SUBSETS from all data (proportion and latency) by picking only one ES from the curves/tests with >1 scale points (we use the prepared columns that already code these choices):

dat_MID_ES <- dplyr::filter(dat, ScalePoint == "MID") #subset by picking only middle (MID) ES from each curve
dim(dat_MID_ES) 

dat_max_absES <- dplyr::filter(dat, biggest_ES == 1) #subset by picking ES with the biggest absolute value from each curve (i.e. get the most positive or negatve response available)
dim(dat_max_absES) #108

dat_max_meanES <- dplyr::filter(dat, biggest_meandir == 1) #subset by picking ES from each curve which has the biggest absolute value in the direction of the mean value of all ES from the curve (i.e. biggest in the overall direction of the response)
dim(dat_max_meanES) #108

dat_max_ES <- dplyr::filter(dat, max_ES == 1) #subset by picking maximum ES from each curve (i.e. get the most positive=optimistic response available)
dim(dat_max_ES) #108
```

###################################################################
#### Table S4   
MA with VCV and phylo - all data and 5 data subsets

```{r Table MA phylo}
#list.files("./Rdata/")
# loading data
load("./Rdata/MA_all.Rdata")
load("./Rdata/MA_amb.Rdata")
load("./Rdata/MA_MID_ES.Rdata")
load("./Rdata/MA_maxabsES.Rdata")
load("./Rdata/MA_maxmeanES.Rdata")
load("./Rdata/MA_maxES.Rdata")

# calculate I2 values for each MA:
MA_all_I2 <- I2(MA_all, method = "Shinichi")
MA_amb_I2 <- I2(MA_amb, method = "Shinichi")
MA_MID_ES_I2 <- I2(MA_MID_ES, method = "Shinichi")
MA_maxabsES_I2 <- I2(MA_maxabsES, method = "Shinichi")
MA_maxmeanES_I2 <- I2(MA_maxmeanES, method = "Shinichi")
MA_maxES_I2 <- I2(MA_maxES, method = "Shinichi")

t4 <- read_excel("./tables/Results_Tables_v3.xlsx", sheet = 1) # loading table for updating MA results

pos <- c(3,10,17,24,31,38) # saving initial row position for each model

# adding results from the models to the table
t4$`Fixed effects`[pos] <- round(c(MA_all$b,
                                   MA_amb$b,
                                   MA_MID_ES$b,
                                   MA_maxabsES$b,
                                   MA_maxmeanES$b,
                                   MA_maxES$b), 3)

t4$...4[pos] <- round(c(MA_all$ci.lb,
                                   MA_amb$ci.lb,
                                   MA_MID_ES$ci.lb,
                                   MA_maxabsES$ci.lb,
                                   MA_maxmeanES$ci.lb,
                                   MA_maxES$ci.lb), 3)

t4$...5[pos] <- round(c(MA_all$ci.ub,
                                   MA_amb$ci.ub,
                                   MA_MID_ES$ci.ub,
                                   MA_maxabsES$ci.ub,
                                   MA_maxmeanES$ci.ub,
                                   MA_maxES$ci.ub), 3)

t4$`Random effects`[pos] <- round(c(MA_all_I2[1]*100,
                                   MA_amb_I2[1]*100,
                                   MA_MID_ES_I2[1]*100,
                                   MA_maxabsES_I2[1]*100,
                                   MA_maxmeanES_I2[1]*100,
                                   MA_maxES_I2[1]*100), 1)

t4$`Random effects`[pos + 1] <- round(c(MA_all_I2[2]*100,
                                   MA_amb_I2[2]*100,
                                   MA_MID_ES_I2[2]*100,
                                   MA_maxabsES_I2[2]*100,
                                   MA_maxmeanES_I2[2]*100,
                                   MA_maxES_I2[2]*100), 1)

t4$`Random effects`[pos + 2] <- round(c(MA_all_I2[3]*100,
                                   MA_amb_I2[3]*100,
                                   MA_MID_ES_I2[3]*100,
                                   MA_maxabsES_I2[3]*100,
                                   MA_maxmeanES_I2[3]*100,
                                   MA_maxES_I2[3]*100), 1)

t4$`Random effects`[pos + 3] <- round(c(MA_all_I2[4]*100,
                                   MA_amb_I2[4]*100,
                                   MA_MID_ES_I2[4]*100,
                                   MA_maxabsES_I2[4]*100,
                                   MA_maxmeanES_I2[4]*100,
                                   MA_maxES_I2[4]*100), 1)

t4$`Random effects`[pos + 4] <- round(c(MA_all_I2[5]*100,
                                   MA_amb_I2[5]*100,
                                   MA_MID_ES_I2[5]*100,
                                   MA_maxabsES_I2[5]*100,
                                   MA_maxmeanES_I2[5]*100,
                                   MA_maxES_I2[5]*100), 1)

t4$`Random effects`[pos + 5] <- round(c(MA_all_I2[6]*100,
                                   MA_amb_I2[6]*100,
                                   MA_MID_ES_I2[6]*100,
                                   MA_maxabsES_I2[6]*100,
                                   MA_maxmeanES_I2[6]*100,
                                   MA_maxES_I2[6]*100), 1)

t4$...8[pos] <- round(c(MA_all$k.all,
                                   MA_amb$k.all,
                                   MA_MID_ES$k.all,
                                   MA_maxabsES$k.all,
                                   MA_maxmeanES$k.all,
                                   MA_maxES$k.all), 3)

t4$...8[pos+1] <- round(c(MA_all$s.nlevels[1],
                                   MA_amb$s.nlevels[1],
                                   MA_MID_ES$s.nlevels[1],
                                   MA_maxabsES$s.nlevels[1],
                                   MA_maxmeanES$s.nlevels[1],
                                   MA_maxES$s.nlevels[1]), 3)

t4$...8[pos+2] <- round(c(MA_all$s.nlevels[2],
                                   MA_amb$s.nlevels[2],
                                   MA_MID_ES$s.nlevels[2],
                                   MA_maxabsES$s.nlevels[2],
                                   MA_maxmeanES$s.nlevels[2],
                                   MA_maxES$s.nlevels[2]), 3)

t4$...8[pos+3] <- round(c(MA_all$s.nlevels[3],
                                   MA_amb$s.nlevels[3],
                                   MA_MID_ES$s.nlevels[3],
                                   MA_maxabsES$s.nlevels[3],
                                   MA_maxmeanES$s.nlevels[3],
                                   MA_maxES$s.nlevels[3]), 3)

t4$...8[pos+4] <- round(c(MA_all$s.nlevels[4],
                                   MA_amb$s.nlevels[4],
                                   MA_MID_ES$s.nlevels[4],
                                   MA_maxabsES$s.nlevels[4],
                                   MA_maxmeanES$s.nlevels[4],
                                   MA_maxES$s.nlevels[4]), 3)

t4$...8[pos+5] <- round(c(MA_all$s.nlevels[5],
                                   MA_amb$s.nlevels[5],
                                   MA_MID_ES$s.nlevels[5],
                                   MA_maxabsES$s.nlevels[5],
                                   MA_maxmeanES$s.nlevels[5],
                                   MA_maxES$s.nlevels[5]), 3)

write.csv(t4, file = "./tables/Table_S4_MA.csv", row.names = F, na = "")
```


###################################################################
#### Table S5   
MA with VCV only (no phylo), Species ID used instead - all data and 5 data subsets

```{r Table MA nophylo}
#list.files("./Rdata/")
# loading data
load("./Rdata/MA_all_random_VCV.Rdata")
load("./Rdata/MA_amb_random_VCV.Rdata")
load("./Rdata/MA_MID_ES_random_VCV.Rdata")
load("./Rdata/MA_maxabsES_random_VCV.Rdata")
load("./Rdata/MA_maxmeanES_random_VCV.Rdata")
load("./Rdata/MA_maxES_random_VCV.Rdata")

# calculate I2 values for each MA:
MA_all_random_VCV_I2 <- I2(MA_all_random_VCV, method = "Shinichi")
MA_amb_random_VCV_I2 <- I2(MA_amb_random_VCV, method = "Shinichi")
MA_MID_ES_random_VCV_I2 <- I2(MA_MID_ES_random_VCV, method = "Shinichi")
MA_maxabsES_random_VCV_I2 <- I2(MA_maxabsES_random_VCV, method = "Shinichi")
MA_maxmeanES_random_VCV_I2 <- I2(MA_maxmeanES_random_VCV, method = "Shinichi")
MA_maxES_random_VCV_I2 <- I2(MA_maxES_random_VCV, method = "Shinichi")

t5 <- read_excel("./tables/Results_Tables_v3.xlsx", sheet = 2) # loading table for updating MA results

pos <- c(3,10,17,24,31,38) # saving initial row position for each model

# adding results from the models to the table
t5$`Fixed effects`[pos] <- round(c(MA_all_random_VCV$b,
                                   MA_amb_random_VCV$b,
                                   MA_MID_ES_random_VCV$b,
                                   MA_maxabsES_random_VCV$b,
                                   MA_maxmeanES_random_VCV$b,
                                   MA_maxES_random_VCV$b), 3)

t5$...4[pos] <- round(c(MA_all_random_VCV$ci.lb,
                                   MA_amb_random_VCV$ci.lb,
                                   MA_MID_ES_random_VCV$ci.lb,
                                   MA_maxabsES_random_VCV$ci.lb,
                                   MA_maxmeanES_random_VCV$ci.lb,
                                   MA_maxES_random_VCV$ci.lb), 3)

t5$...5[pos] <- round(c(MA_all_random_VCV$ci.ub,
                                   MA_amb_random_VCV$ci.ub,
                                   MA_MID_ES_random_VCV$ci.ub,
                                   MA_maxabsES_random_VCV$ci.ub,
                                   MA_maxmeanES_random_VCV$ci.ub,
                                   MA_maxES_random_VCV$ci.ub), 3)

t5$`Random effects`[pos] <- round(c(MA_all_random_VCV_I2[1]*100,
                                   MA_amb_random_VCV_I2[1]*100,
                                   MA_MID_ES_random_VCV_I2[1]*100,
                                   MA_maxabsES_random_VCV_I2[1]*100,
                                   MA_maxmeanES_random_VCV_I2[1]*100,
                                   MA_maxES_random_VCV_I2[1]*100), 1)

t5$`Random effects`[pos + 1] <- round(c(MA_all_random_VCV_I2[2]*100,
                                   MA_amb_random_VCV_I2[2]*100,
                                   MA_MID_ES_random_VCV_I2[2]*100,
                                   MA_maxabsES_random_VCV_I2[2]*100,
                                   MA_maxmeanES_random_VCV_I2[2]*100,
                                   MA_maxES_random_VCV_I2[2]*100), 1)

t5$`Random effects`[pos + 2] <- round(c(MA_all_random_VCV_I2[3]*100,
                                   MA_amb_random_VCV_I2[3]*100,
                                   MA_MID_ES_random_VCV_I2[3]*100,
                                   MA_maxabsES_random_VCV_I2[3]*100,
                                   MA_maxmeanES_random_VCV_I2[3]*100,
                                   MA_maxES_random_VCV_I2[3]*100), 1)

t5$`Random effects`[pos + 3] <- round(c(MA_all_random_VCV_I2[4]*100,
                                   MA_amb_random_VCV_I2[4]*100,
                                   MA_MID_ES_random_VCV_I2[4]*100,
                                   MA_maxabsES_random_VCV_I2[4]*100,
                                   MA_maxmeanES_random_VCV_I2[4]*100,
                                   MA_maxES_random_VCV_I2[4]*100), 1)

t5$`Random effects`[pos + 4] <- round(c(MA_all_random_VCV_I2[5]*100,
                                   MA_amb_random_VCV_I2[5]*100,
                                   MA_MID_ES_random_VCV_I2[5]*100,
                                   MA_maxabsES_random_VCV_I2[5]*100,
                                   MA_maxmeanES_random_VCV_I2[5]*100,
                                   MA_maxES_random_VCV_I2[5]*100), 1)

t5$`Random effects`[pos + 5] <- round(c(MA_all_random_VCV_I2[6]*100,
                                   MA_amb_random_VCV_I2[6]*100,
                                   MA_MID_ES_random_VCV_I2[6]*100,
                                   MA_maxabsES_random_VCV_I2[6]*100,
                                   MA_maxmeanES_random_VCV_I2[6]*100,
                                   MA_maxES_random_VCV_I2[6]*100), 1)

t5$...8[pos] <- round(c(MA_all_random_VCV$k.all,
                                   MA_amb_random_VCV$k.all,
                                   MA_MID_ES_random_VCV$k.all,
                                   MA_maxabsES_random_VCV$k.all,
                                   MA_maxmeanES_random_VCV$k.all,
                                   MA_maxES_random_VCV$k.all), 3)

t5$...8[pos+1] <- round(c(MA_all_random_VCV$s.nlevels[1],
                                   MA_amb_random_VCV$s.nlevels[1],
                                   MA_MID_ES_random_VCV$s.nlevels[1],
                                   MA_maxabsES_random_VCV$s.nlevels[1],
                                   MA_maxmeanES_random_VCV$s.nlevels[1],
                                   MA_maxES_random_VCV$s.nlevels[1]), 3)

t5$...8[pos+2] <- round(c(MA_all_random_VCV$s.nlevels[2],
                                   MA_amb_random_VCV$s.nlevels[2],
                                   MA_MID_ES_random_VCV$s.nlevels[2],
                                   MA_maxabsES_random_VCV$s.nlevels[2],
                                   MA_maxmeanES_random_VCV$s.nlevels[2],
                                   MA_maxES_random_VCV$s.nlevels[2]), 3)

t5$...8[pos+3] <- round(c(MA_all_random_VCV$s.nlevels[3],
                                   MA_amb_random_VCV$s.nlevels[3],
                                   MA_MID_ES_random_VCV$s.nlevels[3],
                                   MA_maxabsES_random_VCV$s.nlevels[3],
                                   MA_maxmeanES_random_VCV$s.nlevels[3],
                                   MA_maxES_random_VCV$s.nlevels[3]), 3)

t5$...8[pos+4] <- round(c(MA_all_random_VCV$s.nlevels[4],
                                   MA_amb_random_VCV$s.nlevels[4],
                                   MA_MID_ES_random_VCV$s.nlevels[4],
                                   MA_maxabsES_random_VCV$s.nlevels[4],
                                   MA_maxmeanES_random_VCV$s.nlevels[4],
                                   MA_maxES_random_VCV$s.nlevels[4]), 3)

t5$...8[pos+5] <- round(c(MA_all_random_VCV$s.nlevels[5],
                                   MA_amb_random_VCV$s.nlevels[5],
                                   MA_MID_ES_random_VCV$s.nlevels[5],
                                   MA_maxabsES_random_VCV$s.nlevels[5],
                                   MA_maxmeanES_random_VCV$s.nlevels[5],
                                   MA_maxES_random_VCV$s.nlevels[5]), 3)

write.csv(t5, file = "./tables/Table_S5_MA_nophylo.csv", row.names = F, na = "")
```



###################################################################    
#### Table S6      
MR - univariate Species model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Species}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_species.Rdata")
load("./Rdata/MR_amb_species.Rdata")
load("./Rdata/MR_MID_species.Rdata")
load("./Rdata/MR_maxabsES_species.Rdata")
load("./Rdata/MR_maxmeanES_species.Rdata")
load("./Rdata/MR_maxES_species.Rdata")


# loading table
t6 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 3)

# saving row position for species estimates
pos <- c(2:23, 25:46, 48:69, 71:92, 94:115, 117:138)
# adding species estimates to the table
t6$`Species name`[pos] <- gsub("ScientificName","",c(rownames(summary(MR_all_species)$b), rownames(summary(MR_amb_species)$b), rownames(summary(MR_MID_species)$b), rownames(summary(MR_maxabsES_species)$b), rownames(summary(MR_maxmeanES_species)$b), rownames(summary(MR_maxES_species)$b)))
t6$`Mean`[pos] <- round(c(MR_all_species$b, MR_amb_species$b, MR_MID_species$b, MR_maxabsES_species$b, MR_maxmeanES_species$b, MR_maxES_species$b),3)
t6$`CI.lb`[pos] <- round(c(MR_all_species$ci.lb, MR_amb_species$ci.lb, MR_MID_species$ci.lb, MR_maxabsES_species$ci.lb, MR_maxmeanES_species$ci.lb, MR_maxES_species$ci.lb), 3)
t6$`CI.ub`[pos] <- round(c(MR_all_species$ci.ub, MR_amb_species$ci.ub, MR_MID_species$ci.ub, MR_maxabsES_species$ci.ub, MR_maxmeanES_species$ci.ub, MR_maxES_species$ci.ub), 3)

t6$R2[pos] <- c(
  table(dat$Species_Latin), 
  table(dat_amb$Species_Latin), 
  table(dat_MID_ES$Species_Latin), 
  table(dat_max_absES$Species_Latin), 
  table(dat_max_meanES$Species_Latin), 
  table(dat_max_ES$Species_Latin))

# adding model R2 estimates to the table
posR2 <- c(1, 24, 47, 70, 93, 116) #row positions
t6$`R2`[posR2] <- round(c(R2(MR_all_species)[1], R2(MR_amb_species)[1], R2(MR_MID_species)[1], R2(MR_maxabsES_species)[1], R2(MR_maxmeanES_species)[1], R2(MR_maxES_species)[1]),3)

write.csv(t6, file = "./Tables/Table_S6_MR_species.csv", row.names = F, na = "")
```


###################################################################    
#### Table S7      
MR - univariate Sex model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Sex}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_sex1.Rdata")
load("./Rdata/MR_all_sex.Rdata") 
load("./Rdata/MR_all_sex2.Rdata") 
load("./Rdata/MR_amb_sex1.Rdata")
load("./Rdata/MR_amb_sex.Rdata")
load("./Rdata/MR_amb_sex2.Rdata")
load("./Rdata/MR_MID_sex1.Rdata")
load("./Rdata/MR_MID_sex.Rdata")
load("./Rdata/MR_MID_sex2.Rdata")
load("./Rdata/MR_maxabsES_sex1.Rdata")
load("./Rdata/MR_maxabsES_sex.Rdata")
load("./Rdata/MR_maxabsES_sex2.Rdata")
load("./Rdata/MR_maxmeanES_sex1.Rdata")
load("./Rdata/MR_maxmeanES_sex.Rdata")
load("./Rdata/MR_maxmeanES_sex2.Rdata")
load("./Rdata/MR_maxES_sex1.Rdata")
load("./Rdata/MR_maxES_sex.Rdata")
load("./Rdata/MR_maxES_sex2.Rdata")

# loading table
t7 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 4)

# saving row position for species estimates
pos <- c(2:7, 9:14, 16:21, 23:28, 30:35, 37:42)

t7$`Mean`[pos] <- round(c(
  MR_all_sex1$b, 
  MR_all_sex$b[2:3], 
  MR_all_sex2$b[3], 
  MR_amb_sex1$b, 
  MR_amb_sex$b[2:3], 
  MR_amb_sex2$b[3], 
  MR_MID_sex1$b, 
  MR_MID_sex$b[2:3], 
  MR_MID_sex2$b[3], 
  MR_maxabsES_sex1$b, 
  MR_maxabsES_sex$b[2:3], 
  MR_maxabsES_sex2$b[3], 
  MR_maxmeanES_sex1$b, 
  MR_maxmeanES_sex$b[2:3], 
  MR_maxmeanES_sex2$b[3], 
  MR_maxES_sex1$b, 
  MR_maxES_sex$b[2:3], 
  MR_maxES_sex2$b[3]), 3) 
  
t7$CI.lb[pos] <- round(c(
  MR_all_sex1$ci.lb, 
  MR_all_sex$ci.lb[2:3], 
  MR_all_sex2$ci.lb[3], 
  MR_amb_sex1$ci.lb, 
  MR_amb_sex$ci.lb[2:3], 
  MR_amb_sex2$ci.lb[3], 
  MR_MID_sex1$ci.lb, 
  MR_MID_sex$ci.lb[2:3], 
  MR_MID_sex2$ci.lb[3], 
  MR_maxabsES_sex1$ci.lb, 
  MR_maxabsES_sex$ci.lb[2:3], 
  MR_maxabsES_sex2$ci.lb[3], 
  MR_maxmeanES_sex1$ci.lb, 
  MR_maxmeanES_sex$ci.lb[2:3], 
  MR_maxmeanES_sex2$ci.lb[3], 
  MR_maxES_sex1$ci.lb, 
  MR_maxES_sex$ci.lb[2:3], 
  MR_maxES_sex2$ci.lb[3]), 3) 

t7$CI.ub[pos] <- round(c(
  MR_all_sex1$ci.ub, 
  MR_all_sex$ci.ub[2:3], 
  MR_all_sex2$ci.ub[3], 
  MR_amb_sex1$ci.ub, 
  MR_amb_sex$ci.ub[2:3], 
  MR_amb_sex2$ci.ub[3], 
  MR_MID_sex1$ci.ub, 
  MR_MID_sex$ci.ub[2:3], 
  MR_MID_sex2$ci.ub[3], 
  MR_maxabsES_sex1$ci.ub, 
  MR_maxabsES_sex$ci.ub[2:3], 
  MR_maxabsES_sex2$ci.ub[3], 
  MR_maxmeanES_sex1$ci.ub, 
  MR_maxmeanES_sex$ci.ub[2:3], 
  MR_maxmeanES_sex2$ci.ub[3], 
  MR_maxES_sex1$ci.ub, 
  MR_maxES_sex$ci.ub[2:3], 
  MR_maxES_sex2$ci.ub[3]), 3) 

t7$R2[pos] <- c(
  table(dat$Sex), 
  rep(NA, 3),
  table(dat_amb$Sex), 
  rep(NA, 3),
  table(dat_MID_ES$Sex), 
  rep(NA, 3),
  table(dat_max_absES$Sex), 
  rep(NA, 3),
  table(dat_max_meanES$Sex), 
  rep(NA, 3),
  table(dat_max_ES$Sex), 
  rep(NA, 3)) 

# adding model R2 estimates to the table
posR2 <- c(1, 8, 15, 22, 29, 36) #row positions
t7$`R2`[posR2] <- round(c(R2(MR_all_sex)[1], R2(MR_amb_sex)[1], R2(MR_MID_sex)[1], R2(MR_maxabsES_sex)[1], R2(MR_maxmeanES_sex)[1], R2(MR_maxES_sex)[1]),3)

write.csv(t7, file = "./Tables/Table_S7_MR_Sex.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S8      
MR - univariate TaskType model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ taskType}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_TaskType1.Rdata")
load("./Rdata/MR_all_TaskType.Rdata") 
load("./Rdata/MR_amb_TaskType1.Rdata")
load("./Rdata/MR_amb_TaskType.Rdata")
load("./Rdata/MR_MID_TaskType1.Rdata")
load("./Rdata/MR_MID_TaskType.Rdata")
load("./Rdata/MR_maxabsES_TaskType1.Rdata")
load("./Rdata/MR_maxabsES_TaskType.Rdata")
load("./Rdata/MR_maxmeanES_TaskType1.Rdata")
load("./Rdata/MR_maxmeanES_TaskType.Rdata")
load("./Rdata/MR_maxES_TaskType1.Rdata")
load("./Rdata/MR_maxES_TaskType.Rdata")

# loading table
t8 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 5)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t8$`Mean`[pos] <- round(c(
  MR_all_TaskType1$b, 
  MR_all_TaskType$b[2], 
  MR_amb_TaskType1$b, 
  MR_amb_TaskType$b[2], 
  MR_MID_TaskType1$b, 
  MR_MID_TaskType$b[2], 
  MR_maxabsES_TaskType1$b, 
  MR_maxabsES_TaskType$b[2], 
  MR_maxmeanES_TaskType1$b, 
  MR_maxmeanES_TaskType$b[2], 
  MR_maxES_TaskType1$b, 
  MR_maxES_TaskType$b[2]), 3) 
  
t8$CI.lb[pos] <- round(c(
  MR_all_TaskType1$ci.lb, 
  MR_all_TaskType$ci.lb[2], 
  MR_amb_TaskType1$ci.lb, 
  MR_amb_TaskType$ci.lb[2], 
  MR_MID_TaskType1$ci.lb, 
  MR_MID_TaskType$ci.lb[2], 
  MR_maxabsES_TaskType1$ci.lb, 
  MR_maxabsES_TaskType$ci.lb[2], 
  MR_maxmeanES_TaskType1$ci.lb, 
  MR_maxmeanES_TaskType$ci.lb[2], 
  MR_maxES_TaskType1$ci.lb, 
  MR_maxES_TaskType$ci.lb[2]), 3) 

t8$CI.ub[pos] <- round(c(
  MR_all_TaskType1$ci.ub, 
  MR_all_TaskType$ci.ub[2], 
  MR_amb_TaskType1$ci.ub, 
  MR_amb_TaskType$ci.ub[2], 
  MR_MID_TaskType1$ci.ub, 
  MR_MID_TaskType$ci.ub[2], 
  MR_maxabsES_TaskType1$ci.ub, 
  MR_maxabsES_TaskType$ci.ub[2], 
  MR_maxmeanES_TaskType1$ci.ub, 
  MR_maxmeanES_TaskType$ci.ub[2], 
  MR_maxES_TaskType1$ci.ub, 
  MR_maxES_TaskType$ci.ub[2]), 3) 

t8$R2[pos] <- c(
  table(dat$TaskType), 
  rep(NA, 1),
  table(dat_amb$TaskType), 
  rep(NA, 1),
  table(dat_MID_ES$TaskType), 
  rep(NA, 1),
  table(dat_max_absES$TaskType), 
  rep(NA, 1),
  table(dat_max_meanES$TaskType), 
  rep(NA, 1),
  table(dat_max_ES$TaskType), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t8$`R2`[posR2] <- round(c(R2(MR_all_TaskType)[1], R2(MR_amb_TaskType)[1], R2(MR_MID_TaskType)[1], R2(MR_maxabsES_TaskType)[1], R2(MR_maxmeanES_TaskType)[1], R2(MR_maxES_TaskType)[1]),3)

write.csv(t8, file = "./Tables/Table_S8_MR_TaskType.csv", row.names = F, na = "")    
```



###################################################################    
#### Table S9      
MR - univariate CueTypeCat model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ CueTypeCat}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_CueTypeCat1.Rdata")
load("./Rdata/MR_all_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_all_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_all_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_all_CueTypeCat_tactile.Rdata") 

load("./Rdata/MR_amb_CueTypeCat1.Rdata")
load("./Rdata/MR_amb_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_amb_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_amb_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_amb_CueTypeCat_tactile.Rdata")

load("./Rdata/MR_MID_CueTypeCat1.Rdata")
load("./Rdata/MR_MID_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_MID_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_MID_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_MID_CueTypeCat_tactile.Rdata")

load("./Rdata/MR_maxabsES_CueTypeCat1.Rdata")
load("./Rdata/MR_maxabsES_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_maxabsES_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_maxabsES_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_maxabsES_CueTypeCat_tactile.Rdata")

load("./Rdata/MR_maxmeanES_CueTypeCat1.Rdata")
load("./Rdata/MR_maxmeanES_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_maxmeanES_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_maxmeanES_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_maxmeanES_CueTypeCat_tactile.Rdata")

load("./Rdata/MR_maxES_CueTypeCat1.Rdata")
load("./Rdata/MR_maxES_CueTypeCat_auditory.Rdata") 
load("./Rdata/MR_maxES_CueTypeCat_olfactory.Rdata") 
load("./Rdata/MR_maxES_CueTypeCat_spatial.Rdata") 
load("./Rdata/MR_maxES_CueTypeCat_tactile.Rdata")

# loading table
t9 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 6)

# saving row position for species estimates
pos <- c(2:16, 18:32, 34:48, 50:64, 66:80, 82:96)

t9$`Mean`[pos] <- round(c(
  MR_all_CueTypeCat1$b, 
  MR_all_CueTypeCat_auditory$b[2:5], 
  MR_all_CueTypeCat_olfactory$b[3:5], 
  MR_all_CueTypeCat_spatial$b[4:5], 
  MR_all_CueTypeCat_tactile$b[5], 
  
  MR_amb_CueTypeCat1$b, 
  MR_amb_CueTypeCat_auditory$b[2:5], 
  MR_amb_CueTypeCat_olfactory$b[3:5], 
  MR_amb_CueTypeCat_spatial$b[4:5], 
  MR_amb_CueTypeCat_tactile$b[5], 
  
  MR_MID_CueTypeCat1$b, 
  MR_MID_CueTypeCat_auditory$b[2:5], 
  MR_MID_CueTypeCat_olfactory$b[3:5], 
  MR_MID_CueTypeCat_spatial$b[4:5], 
  MR_MID_CueTypeCat_tactile$b[5], 
  
  MR_maxabsES_CueTypeCat1$b, 
  MR_maxabsES_CueTypeCat_auditory$b[2:5], 
  MR_maxabsES_CueTypeCat_olfactory$b[3:5], 
  MR_maxabsES_CueTypeCat_spatial$b[4:5], 
  MR_maxabsES_CueTypeCat_tactile$b[5], 
  
  MR_maxmeanES_CueTypeCat1$b, 
  MR_maxmeanES_CueTypeCat_auditory$b[2:5], 
  MR_maxmeanES_CueTypeCat_olfactory$b[3:5], 
  MR_maxmeanES_CueTypeCat_spatial$b[4:5], 
  MR_maxmeanES_CueTypeCat_tactile$b[5], 
  
  MR_maxES_CueTypeCat1$b, 
  MR_maxES_CueTypeCat_auditory$b[2:5], 
  MR_maxES_CueTypeCat_olfactory$b[3:5], 
  MR_maxES_CueTypeCat_spatial$b[4:5], 
  MR_maxES_CueTypeCat_tactile$b[5]), 3) 
  
t9$CI.lb[pos] <- round(c(
  MR_all_CueTypeCat1$ci.lb, 
  MR_all_CueTypeCat_auditory$ci.lb[2:5], 
  MR_all_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_all_CueTypeCat_spatial$ci.lb[4:5], 
  MR_all_CueTypeCat_tactile$ci.lb[5], 
  
  MR_amb_CueTypeCat1$ci.lb, 
  MR_amb_CueTypeCat_auditory$ci.lb[2:5], 
  MR_amb_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_amb_CueTypeCat_spatial$ci.lb[4:5], 
  MR_amb_CueTypeCat_tactile$ci.lb[5], 
  
  MR_MID_CueTypeCat1$ci.lb, 
  MR_MID_CueTypeCat_auditory$ci.lb[2:5], 
  MR_MID_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_MID_CueTypeCat_spatial$ci.lb[4:5], 
  MR_MID_CueTypeCat_tactile$ci.lb[5], 
  
  MR_maxabsES_CueTypeCat1$ci.lb, 
  MR_maxabsES_CueTypeCat_auditory$ci.lb[2:5], 
  MR_maxabsES_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_maxabsES_CueTypeCat_spatial$ci.lb[4:5], 
  MR_maxabsES_CueTypeCat_tactile$ci.lb[5], 
  
  MR_maxmeanES_CueTypeCat1$ci.lb, 
  MR_maxmeanES_CueTypeCat_auditory$ci.lb[2:5], 
  MR_maxmeanES_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_maxmeanES_CueTypeCat_spatial$ci.lb[4:5], 
  MR_maxmeanES_CueTypeCat_tactile$ci.lb[5], 
  
  MR_maxES_CueTypeCat1$ci.lb, 
  MR_maxES_CueTypeCat_auditory$ci.lb[2:5], 
  MR_maxES_CueTypeCat_olfactory$ci.lb[3:5], 
  MR_maxES_CueTypeCat_spatial$ci.lb[4:5], 
  MR_maxES_CueTypeCat_tactile$ci.lb[5]), 3) 

t9$CI.ub[pos] <- round(c(
  MR_all_CueTypeCat1$ci.ub, 
  MR_all_CueTypeCat_auditory$ci.ub[2:5], 
  MR_all_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_all_CueTypeCat_spatial$ci.ub[4:5], 
  MR_all_CueTypeCat_tactile$ci.ub[5], 
  
  MR_amb_CueTypeCat1$ci.ub, 
  MR_amb_CueTypeCat_auditory$ci.ub[2:5], 
  MR_amb_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_amb_CueTypeCat_spatial$ci.ub[4:5], 
  MR_amb_CueTypeCat_tactile$ci.ub[5], 
  
  MR_MID_CueTypeCat1$ci.ub, 
  MR_MID_CueTypeCat_auditory$ci.ub[2:5], 
  MR_MID_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_MID_CueTypeCat_spatial$ci.ub[4:5], 
  MR_MID_CueTypeCat_tactile$ci.ub[5], 
  
  MR_maxabsES_CueTypeCat1$ci.ub, 
  MR_maxabsES_CueTypeCat_auditory$ci.ub[2:5], 
  MR_maxabsES_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_maxabsES_CueTypeCat_spatial$ci.ub[4:5], 
  MR_maxabsES_CueTypeCat_tactile$ci.ub[5], 
  
  MR_maxmeanES_CueTypeCat1$ci.ub, 
  MR_maxmeanES_CueTypeCat_auditory$ci.ub[2:5], 
  MR_maxmeanES_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_maxmeanES_CueTypeCat_spatial$ci.ub[4:5], 
  MR_maxmeanES_CueTypeCat_tactile$ci.ub[5], 
  
  MR_maxES_CueTypeCat1$ci.ub, 
  MR_maxES_CueTypeCat_auditory$ci.ub[2:5], 
  MR_maxES_CueTypeCat_olfactory$ci.ub[3:5], 
  MR_maxES_CueTypeCat_spatial$ci.ub[4:5], 
  MR_maxES_CueTypeCat_tactile$ci.ub[5]), 3)  

t9$R2[pos] <- c(
  table(dat$CueTypeCat), 
  rep(NA, 10),
  table(dat_amb$CueTypeCat), 
  rep(NA, 10),
  table(dat_MID_ES$CueTypeCat), 
  rep(NA, 10),
  table(dat_max_absES$CueTypeCat), 
  rep(NA, 10),
  table(dat_max_meanES$CueTypeCat), 
  rep(NA, 10),
  table(dat_max_ES$CueTypeCat), 
  rep(NA, 10)) 

# adding model R2 estimates to the table
posR2 <- c(1, 17, 33, 49, 65, 81) #row positions
t9$`R2`[posR2] <- round(c(R2(MR_all_CueTypeCat1)[1], R2(MR_amb_CueTypeCat1)[1], R2(MR_MID_CueTypeCat1)[1], R2(MR_maxabsES_CueTypeCat1)[1], R2(MR_maxmeanES_CueTypeCat1)[1], R2(MR_maxES_CueTypeCat1)[1]),3)

write.csv(t9, file = "./Tables/Table_S9_MR_CueTypeCat.csv", row.names = F, na = "")    
```




###################################################################    
#### Table S10      
MR - univariate ReinforcementCat model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ ReinforcementCat}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_ReinforcementCat1.Rdata")
load("./Rdata/MR_all_ReinforcementCat.Rdata") 
load("./Rdata/MR_all_ReinforcementCat2.Rdata") 
load("./Rdata/MR_amb_ReinforcementCat1.Rdata")
load("./Rdata/MR_amb_ReinforcementCat.Rdata")
load("./Rdata/MR_amb_ReinforcementCat2.Rdata")
load("./Rdata/MR_MID_ReinforcementCat1.Rdata")
load("./Rdata/MR_MID_ReinforcementCat.Rdata")
load("./Rdata/MR_MID_ReinforcementCat2.Rdata")
load("./Rdata/MR_maxabsES_ReinforcementCat1.Rdata")
load("./Rdata/MR_maxabsES_ReinforcementCat.Rdata")
load("./Rdata/MR_maxabsES_ReinforcementCat2.Rdata")
load("./Rdata/MR_maxmeanES_ReinforcementCat1.Rdata")
load("./Rdata/MR_maxmeanES_ReinforcementCat.Rdata")
load("./Rdata/MR_maxmeanES_ReinforcementCat2.Rdata")
load("./Rdata/MR_maxES_ReinforcementCat1.Rdata")
load("./Rdata/MR_maxES_ReinforcementCat.Rdata")
load("./Rdata/MR_maxES_ReinforcementCat2.Rdata")

# loading table
t10 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 7)

# saving row position for species estimates
pos <- c(2:7, 9:14, 16:21, 23:28, 30:35, 37:42)

t10$`Mean`[pos] <- round(c(
  MR_all_ReinforcementCat1$b, 
  MR_all_ReinforcementCat$b[2:3], 
  MR_all_ReinforcementCat2$b[3], 
  MR_amb_ReinforcementCat1$b, 
  MR_amb_ReinforcementCat$b[2:3], 
  MR_amb_ReinforcementCat2$b[3], 
  MR_MID_ReinforcementCat1$b, 
  MR_MID_ReinforcementCat$b[2:3], 
  MR_MID_ReinforcementCat2$b[3], 
  MR_maxabsES_ReinforcementCat1$b, 
  MR_maxabsES_ReinforcementCat$b[2:3], 
  MR_maxabsES_ReinforcementCat2$b[3], 
  MR_maxmeanES_ReinforcementCat1$b, 
  MR_maxmeanES_ReinforcementCat$b[2:3], 
  MR_maxmeanES_ReinforcementCat2$b[3], 
  MR_maxES_ReinforcementCat1$b, 
  MR_maxES_ReinforcementCat$b[2:3], 
  MR_maxES_ReinforcementCat2$b[3]), 3) 
  
t10$CI.lb[pos] <- round(c(
  MR_all_ReinforcementCat1$ci.lb, 
  MR_all_ReinforcementCat$ci.lb[2:3], 
  MR_all_ReinforcementCat2$ci.lb[3], 
  MR_amb_ReinforcementCat1$ci.lb, 
  MR_amb_ReinforcementCat$ci.lb[2:3], 
  MR_amb_ReinforcementCat2$ci.lb[3], 
  MR_MID_ReinforcementCat1$ci.lb, 
  MR_MID_ReinforcementCat$ci.lb[2:3], 
  MR_MID_ReinforcementCat2$ci.lb[3], 
  MR_maxabsES_ReinforcementCat1$ci.lb, 
  MR_maxabsES_ReinforcementCat$ci.lb[2:3], 
  MR_maxabsES_ReinforcementCat2$ci.lb[3], 
  MR_maxmeanES_ReinforcementCat1$ci.lb, 
  MR_maxmeanES_ReinforcementCat$ci.lb[2:3], 
  MR_maxmeanES_ReinforcementCat2$ci.lb[3], 
  MR_maxES_ReinforcementCat1$ci.lb, 
  MR_maxES_ReinforcementCat$ci.lb[2:3], 
  MR_maxES_ReinforcementCat2$ci.lb[3]), 3) 

t10$CI.ub[pos] <- round(c(
  MR_all_ReinforcementCat1$ci.ub, 
  MR_all_ReinforcementCat$ci.ub[2:3], 
  MR_all_ReinforcementCat2$ci.ub[3], 
  MR_amb_ReinforcementCat1$ci.ub, 
  MR_amb_ReinforcementCat$ci.ub[2:3], 
  MR_amb_ReinforcementCat2$ci.ub[3], 
  MR_MID_ReinforcementCat1$ci.ub, 
  MR_MID_ReinforcementCat$ci.ub[2:3], 
  MR_MID_ReinforcementCat2$ci.ub[3], 
  MR_maxabsES_ReinforcementCat1$ci.ub, 
  MR_maxabsES_ReinforcementCat$ci.ub[2:3], 
  MR_maxabsES_ReinforcementCat2$ci.ub[3], 
  MR_maxmeanES_ReinforcementCat1$ci.ub, 
  MR_maxmeanES_ReinforcementCat$ci.ub[2:3], 
  MR_maxmeanES_ReinforcementCat2$ci.ub[3], 
  MR_maxES_ReinforcementCat1$ci.ub, 
  MR_maxES_ReinforcementCat$ci.ub[2:3], 
  MR_maxES_ReinforcementCat2$ci.ub[3]), 3) 

t10$R2[pos] <- c(
  table(dat$ReinforcementCat), 
  rep(NA, 3),
  table(dat_amb$ReinforcementCat), 
  rep(NA, 3),
  table(dat_MID_ES$ReinforcementCat), 
  rep(NA, 3),
  table(dat_max_absES$ReinforcementCat), 
  rep(NA, 3),
  table(dat_max_meanES$ReinforcementCat), 
  rep(NA, 3),
  table(dat_max_ES$ReinforcementCat), 
  rep(NA, 3)) 

# adding model R2 estimates to the table
posR2 <- c(1, 8, 15, 22, 29, 36) #row positions
t10$`R2`[posR2] <- round(c(R2(MR_all_ReinforcementCat)[1], R2(MR_amb_ReinforcementCat)[1], R2(MR_MID_ReinforcementCat)[1], R2(MR_maxabsES_ReinforcementCat)[1], R2(MR_maxmeanES_ReinforcementCat)[1], R2(MR_maxES_ReinforcementCat)[1]),3)

write.csv(t10, file = "./Tables/Table_S10_MR_ReinforcementCat.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S11      
MR - univariate AmbigReinforced model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ AmbigReinforced}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_AmbigReinforced1.Rdata")
load("./Rdata/MR_all_AmbigReinforced.Rdata") 
load("./Rdata/MR_amb_AmbigReinforced1.Rdata")
load("./Rdata/MR_amb_AmbigReinforced.Rdata")
load("./Rdata/MR_MID_AmbigReinforced1.Rdata")
load("./Rdata/MR_MID_AmbigReinforced.Rdata")
load("./Rdata/MR_maxabsES_AmbigReinforced1.Rdata")
load("./Rdata/MR_maxabsES_AmbigReinforced.Rdata")
load("./Rdata/MR_maxmeanES_AmbigReinforced1.Rdata")
load("./Rdata/MR_maxmeanES_AmbigReinforced.Rdata")
load("./Rdata/MR_maxES_AmbigReinforced1.Rdata")
load("./Rdata/MR_maxES_AmbigReinforced.Rdata")

# loading table
t11 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 8)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t11$`Mean`[pos] <- round(c(
  MR_all_AmbigReinforced1$b, 
  MR_all_AmbigReinforced$b[2], 
  MR_amb_AmbigReinforced1$b, 
  MR_amb_AmbigReinforced$b[2], 
  MR_MID_AmbigReinforced1$b, 
  MR_MID_AmbigReinforced$b[2], 
  MR_maxabsES_AmbigReinforced1$b, 
  MR_maxabsES_AmbigReinforced$b[2], 
  MR_maxmeanES_AmbigReinforced1$b, 
  MR_maxmeanES_AmbigReinforced$b[2], 
  MR_maxES_AmbigReinforced1$b, 
  MR_maxES_AmbigReinforced$b[2]), 3) 
  
t11$CI.lb[pos] <- round(c(
  MR_all_AmbigReinforced1$ci.lb, 
  MR_all_AmbigReinforced$ci.lb[2], 
  MR_amb_AmbigReinforced1$ci.lb, 
  MR_amb_AmbigReinforced$ci.lb[2], 
  MR_MID_AmbigReinforced1$ci.lb, 
  MR_MID_AmbigReinforced$ci.lb[2], 
  MR_maxabsES_AmbigReinforced1$ci.lb, 
  MR_maxabsES_AmbigReinforced$ci.lb[2], 
  MR_maxmeanES_AmbigReinforced1$ci.lb, 
  MR_maxmeanES_AmbigReinforced$ci.lb[2], 
  MR_maxES_AmbigReinforced1$ci.lb, 
  MR_maxES_AmbigReinforced$ci.lb[2]), 3) 

t11$CI.ub[pos] <- round(c(
  MR_all_AmbigReinforced1$ci.ub, 
  MR_all_AmbigReinforced$ci.ub[2], 
  MR_amb_AmbigReinforced1$ci.ub, 
  MR_amb_AmbigReinforced$ci.ub[2], 
  MR_MID_AmbigReinforced1$ci.ub, 
  MR_MID_AmbigReinforced$ci.ub[2], 
  MR_maxabsES_AmbigReinforced1$ci.ub, 
  MR_maxabsES_AmbigReinforced$ci.ub[2], 
  MR_maxmeanES_AmbigReinforced1$ci.ub, 
  MR_maxmeanES_AmbigReinforced$ci.ub[2], 
  MR_maxES_AmbigReinforced1$ci.ub, 
  MR_maxES_AmbigReinforced$ci.ub[2]), 3) 

t11$R2[pos] <- c(
  table(dat$AmbigReinforced), 
  rep(NA, 1),
  table(dat_amb$AmbigReinforced), 
  rep(NA, 1),
  table(dat_MID_ES$AmbigReinforced), 
  rep(NA, 1),
  table(dat_max_absES$AmbigReinforced), 
  rep(NA, 1),
  table(dat_max_meanES$AmbigReinforced), 
  rep(NA, 1),
  table(dat_max_ES$AmbigReinforced), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t11$`R2`[posR2] <- round(c(R2(MR_all_AmbigReinforced)[1], R2(MR_amb_AmbigReinforced)[1], R2(MR_MID_AmbigReinforced)[1], R2(MR_maxabsES_AmbigReinforced)[1], R2(MR_maxmeanES_AmbigReinforced)[1], R2(MR_maxES_AmbigReinforced)[1]),3)

write.csv(t11, file = "./Tables/Table_S11_MR_AmbigReinforced.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S12      
MR - univariate ScalePoint model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ ScalePoint}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_ScalePoint1.Rdata")
load("./Rdata/MR_all_ScalePoint_P.Rdata") 
load("./Rdata/MR_all_ScalePoint_NP.Rdata") 
load("./Rdata/MR_all_ScalePoint_MID.Rdata") 
load("./Rdata/MR_all_ScalePoint_NN.Rdata") 

load("./Rdata/MR_amb_ScalePoint1.Rdata")
load("./Rdata/MR_amb_ScalePoint_P.Rdata")
load("./Rdata/MR_amb_ScalePoint_NP.Rdata") 
load("./Rdata/MR_amb_ScalePoint_MID.Rdata") 
load("./Rdata/MR_amb_ScalePoint_NN.Rdata")

load("./Rdata/MR_MID_ScalePoint1.Rdata")
load("./Rdata/MR_MID_ScalePoint_P.Rdata") 
load("./Rdata/MR_MID_ScalePoint_NP.Rdata") 
load("./Rdata/MR_MID_ScalePoint_MID.Rdata") 
load("./Rdata/MR_MID_ScalePoint_NN.Rdata")

load("./Rdata/MR_maxabsES_ScalePoint1.Rdata")
load("./Rdata/MR_maxabsES_ScalePoint_P.Rdata") 
load("./Rdata/MR_maxabsES_ScalePoint_NP.Rdata") 
load("./Rdata/MR_maxabsES_ScalePoint_MID.Rdata") 
load("./Rdata/MR_maxabsES_ScalePoint_NN.Rdata")

load("./Rdata/MR_maxmeanES_ScalePoint1.Rdata")
load("./Rdata/MR_maxmeanES_ScalePoint_P.Rdata") 
load("./Rdata/MR_maxmeanES_ScalePoint_NP.Rdata") 
load("./Rdata/MR_maxmeanES_ScalePoint_MID.Rdata") 
load("./Rdata/MR_maxmeanES_ScalePoint_NN.Rdata")

load("./Rdata/MR_maxES_ScalePoint1.Rdata")
load("./Rdata/MR_maxES_ScalePoint_P.Rdata") 
load("./Rdata/MR_maxES_ScalePoint_NP.Rdata") 
load("./Rdata/MR_maxES_ScalePoint_MID.Rdata") 
load("./Rdata/MR_maxES_ScalePoint_NN.Rdata")

# loading table
t12 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 9)

# saving row position for species estimates
pos <- c(2:16, 18:32, 34:48, 50:64, 66:80, 82:96)

t12$`Mean`[pos] <- round(c(
  MR_all_ScalePoint1$b[c(5,4,1,3,2)], 
  MR_all_ScalePoint_P$b[c(5,2,4,3)], 
  MR_all_ScalePoint_NP$b[c(2,4,3)], 
  MR_all_ScalePoint_MID$b[c(4,3)], 
  MR_all_ScalePoint_NN$b[3], 
  NA,
  MR_amb_ScalePoint1$b[c(3,1,2)], 
  rep(NA,5),
  MR_amb_ScalePoint_NP$b[c(2,3)], 
  NA,
  MR_amb_ScalePoint_MID$b[c(2)],
  rep(NA,4),
  MR_MID_ScalePoint1$b[1], 
  rep(NA,12),
  MR_maxabsES_ScalePoint1$b[c(5,4,1,3,2)], 
  MR_maxabsES_ScalePoint_P$b[c(5,2,4,3)], 
  MR_maxabsES_ScalePoint_NP$b[c(2,4,3)], 
  MR_maxabsES_ScalePoint_MID$b[c(4,3)], 
  MR_maxabsES_ScalePoint_NN$b[3], 
  MR_maxmeanES_ScalePoint1$b[c(5,4,1,3,2)], 
  MR_maxmeanES_ScalePoint_P$b[c(5,2,4,3)], 
  MR_maxmeanES_ScalePoint_NP$b[c(2,4,3)], 
  MR_maxmeanES_ScalePoint_MID$b[c(4,3)], 
  MR_maxmeanES_ScalePoint_NN$b[3], 
  MR_maxES_ScalePoint1$b[c(5,4,1,3,2)], 
  MR_maxES_ScalePoint_P$b[c(5,2,4,3)], 
  MR_maxES_ScalePoint_NP$b[c(2,4,3)], 
  MR_maxES_ScalePoint_MID$b[c(4,3)], 
  MR_maxES_ScalePoint_NN$b[3]), 3) 
  
t12$CI.lb[pos] <- round(c(
  MR_all_ScalePoint1$ci.lb[c(5,4,1,3,2)], 
  MR_all_ScalePoint_P$ci.lb[c(5,2,4,3)], 
  MR_all_ScalePoint_NP$ci.lb[c(2,4,3)], 
  MR_all_ScalePoint_MID$ci.lb[c(4,3)], 
  MR_all_ScalePoint_NN$ci.lb[3], 
  NA,
  MR_amb_ScalePoint1$ci.lb[c(3,1,2)], 
  rep(NA,5),
  MR_amb_ScalePoint_NP$ci.lb[c(2,3)], 
  NA,
  MR_amb_ScalePoint_MID$ci.lb[c(2)],
  rep(NA,4),
  MR_MID_ScalePoint1$ci.lb[1], 
  rep(NA,12),
  MR_maxabsES_ScalePoint1$ci.lb[c(5,4,1,3,2)], 
  MR_maxabsES_ScalePoint_P$ci.lb[c(5,2,4,3)], 
  MR_maxabsES_ScalePoint_NP$ci.lb[c(2,4,3)], 
  MR_maxabsES_ScalePoint_MID$ci.lb[c(4,3)], 
  MR_maxabsES_ScalePoint_NN$ci.lb[3], 
  MR_maxmeanES_ScalePoint1$ci.lb[c(5,4,1,3,2)], 
  MR_maxmeanES_ScalePoint_P$ci.lb[c(5,2,4,3)], 
  MR_maxmeanES_ScalePoint_NP$ci.lb[c(2,4,3)], 
  MR_maxmeanES_ScalePoint_MID$ci.lb[c(4,3)], 
  MR_maxmeanES_ScalePoint_NN$ci.lb[3], 
  MR_maxES_ScalePoint1$ci.lb[c(5,4,1,3,2)], 
  MR_maxES_ScalePoint_P$ci.lb[c(5,2,4,3)], 
  MR_maxES_ScalePoint_NP$ci.lb[c(2,4,3)], 
  MR_maxES_ScalePoint_MID$ci.lb[c(4,3)], 
  MR_maxES_ScalePoint_NN$ci.lb[3]), 3) 

t12$CI.ub[pos] <- round(c(
  MR_all_ScalePoint1$ci.ub[c(5,4,1,3,2)], 
  MR_all_ScalePoint_P$ci.ub[c(5,2,4,3)], 
  MR_all_ScalePoint_NP$ci.ub[c(2,4,3)], 
  MR_all_ScalePoint_MID$ci.ub[c(4,3)], 
  MR_all_ScalePoint_NN$ci.ub[3], 
  NA,
  MR_amb_ScalePoint1$ci.ub[c(3,1,2)], 
  rep(NA,5),
  MR_amb_ScalePoint_NP$ci.ub[c(2,3)], 
  NA,
  MR_amb_ScalePoint_MID$ci.ub[c(2)],
  rep(NA,4),
  MR_MID_ScalePoint1$ci.ub[1], 
  rep(NA,12),
  MR_maxabsES_ScalePoint1$ci.ub[c(5,4,1,3,2)], 
  MR_maxabsES_ScalePoint_P$ci.ub[c(5,2,4,3)], 
  MR_maxabsES_ScalePoint_NP$ci.ub[c(2,4,3)], 
  MR_maxabsES_ScalePoint_MID$ci.ub[c(4,3)], 
  MR_maxabsES_ScalePoint_NN$ci.ub[3], 
  MR_maxmeanES_ScalePoint1$ci.ub[c(5,4,1,3,2)], 
  MR_maxmeanES_ScalePoint_P$ci.ub[c(5,2,4,3)], 
  MR_maxmeanES_ScalePoint_NP$ci.ub[c(2,4,3)], 
  MR_maxmeanES_ScalePoint_MID$ci.ub[c(4,3)], 
  MR_maxmeanES_ScalePoint_NN$ci.ub[3], 
  MR_maxES_ScalePoint1$ci.ub[c(5,4,1,3,2)], 
  MR_maxES_ScalePoint_P$ci.ub[c(5,2,4,3)], 
  MR_maxES_ScalePoint_NP$ci.ub[c(2,4,3)], 
  MR_maxES_ScalePoint_MID$ci.ub[c(4,3)], 
  MR_maxES_ScalePoint_NN$ci.ub[3]), 3) 

t12$R2[pos] <- c(
  table(dat$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10),
  table(dat_amb$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10),
  table(dat_MID_ES$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10),
  table(dat_max_absES$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10),
  table(dat_max_meanES$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10),
  table(dat_max_ES$ScalePoint)[c(5,4,1,3,2)], 
  rep(NA, 10)) 

# adding model R2 estimates to the table
posR2 <- c(1, 17, 33, 49, 65, 81) #row positions
t12$`R2`[posR2] <- round(c(R2(MR_all_ScalePoint1)[1], R2(MR_amb_ScalePoint1)[1], R2(MR_MID_ScalePoint1)[1], R2(MR_maxabsES_ScalePoint1)[1], R2(MR_maxmeanES_ScalePoint1)[1], R2(MR_maxES_ScalePoint1)[1]),3)

write.csv(t12, file = "./Tables/Table_S12_MR_ScalePoint.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S13      
MR - univariate Source (of animals) model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Source}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_source1.Rdata")
load("./Rdata/MR_all_source.Rdata") 
load("./Rdata/MR_amb_source1.Rdata")
load("./Rdata/MR_amb_source.Rdata")
load("./Rdata/MR_MID_source1.Rdata")
load("./Rdata/MR_MID_source.Rdata")
load("./Rdata/MR_maxabsES_source1.Rdata")
load("./Rdata/MR_maxabsES_source.Rdata")
load("./Rdata/MR_maxmeanES_source1.Rdata")
load("./Rdata/MR_maxmeanES_source.Rdata")
load("./Rdata/MR_maxES_source1.Rdata")
load("./Rdata/MR_maxES_source.Rdata")

# loading table
t13 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 10)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t13$`Mean`[pos] <- round(c(
  MR_all_source1$b, 
  MR_all_source$b[2], 
  MR_amb_source1$b, 
  MR_amb_source$b[2], 
  MR_MID_source1$b, 
  MR_MID_source$b[2], 
  MR_maxabsES_source1$b, 
  MR_maxabsES_source$b[2], 
  MR_maxmeanES_source1$b, 
  MR_maxmeanES_source$b[2], 
  MR_maxES_source1$b, 
  MR_maxES_source$b[2]), 3) 
  
t13$CI.lb[pos] <- round(c(
  MR_all_source1$ci.lb, 
  MR_all_source$ci.lb[2], 
  MR_amb_source1$ci.lb, 
  MR_amb_source$ci.lb[2], 
  MR_MID_source1$ci.lb, 
  MR_MID_source$ci.lb[2], 
  MR_maxabsES_source1$ci.lb, 
  MR_maxabsES_source$ci.lb[2], 
  MR_maxmeanES_source1$ci.lb, 
  MR_maxmeanES_source$ci.lb[2], 
  MR_maxES_source1$ci.lb, 
  MR_maxES_source$ci.lb[2]), 3) 

t13$CI.ub[pos] <- round(c(
  MR_all_source1$ci.ub, 
  MR_all_source$ci.ub[2], 
  MR_amb_source1$ci.ub, 
  MR_amb_source$ci.ub[2], 
  MR_MID_source1$ci.ub, 
  MR_MID_source$ci.ub[2], 
  MR_maxabsES_source1$ci.ub, 
  MR_maxabsES_source$ci.ub[2], 
  MR_maxmeanES_source1$ci.ub, 
  MR_maxmeanES_source$ci.ub[2], 
  MR_maxES_source1$ci.ub, 
  MR_maxES_source$ci.ub[2]), 3) 

t13$R2[pos] <- c(
  table(dat$source), 
  rep(NA, 1),
  table(dat_amb$source), 
  rep(NA, 1),
  table(dat_MID_ES$source), 
  rep(NA, 1),
  table(dat_max_absES$source), 
  rep(NA, 1),
  table(dat_max_meanES$source), 
  rep(NA, 1),
  table(dat_max_ES$source), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t13$`R2`[posR2] <- round(c(R2(MR_all_source)[1], R2(MR_amb_source)[1], R2(MR_MID_source)[1], R2(MR_maxabsES_source)[1], R2(MR_maxmeanES_source)[1], R2(MR_maxES_source)[1]),3)

write.csv(t13, file = "./Tables/Table_S13_MR_Source.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S14      
MR - univariate Age (of animals) model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Age}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_age1.Rdata")
load("./Rdata/MR_all_age.Rdata") 
load("./Rdata/MR_amb_age1.Rdata")
load("./Rdata/MR_amb_age.Rdata")
load("./Rdata/MR_MID_age1.Rdata")
load("./Rdata/MR_MID_age.Rdata")
load("./Rdata/MR_maxabsES_age1.Rdata")
load("./Rdata/MR_maxabsES_age.Rdata")
load("./Rdata/MR_maxmeanES_age1.Rdata")
load("./Rdata/MR_maxmeanES_age.Rdata")
load("./Rdata/MR_maxES_age1.Rdata")
load("./Rdata/MR_maxES_age.Rdata")

# loading table
t14 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 11)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t14$`Mean`[pos] <- round(c(
  MR_all_age1$b, 
  MR_all_age$b[2], 
  MR_amb_age1$b, 
  MR_amb_age$b[2], 
  MR_MID_age1$b, 
  MR_MID_age$b[2], 
  MR_maxabsES_age1$b, 
  MR_maxabsES_age$b[2], 
  MR_maxmeanES_age1$b, 
  MR_maxmeanES_age$b[2], 
  MR_maxES_age1$b, 
  MR_maxES_age$b[2]), 3) 
  
t14$CI.lb[pos] <- round(c(
  MR_all_age1$ci.lb, 
  MR_all_age$ci.lb[2], 
  MR_amb_age1$ci.lb, 
  MR_amb_age$ci.lb[2], 
  MR_MID_age1$ci.lb, 
  MR_MID_age$ci.lb[2], 
  MR_maxabsES_age1$ci.lb, 
  MR_maxabsES_age$ci.lb[2], 
  MR_maxmeanES_age1$ci.lb, 
  MR_maxmeanES_age$ci.lb[2], 
  MR_maxES_age1$ci.lb, 
  MR_maxES_age$ci.lb[2]), 3) 

t14$CI.ub[pos] <- round(c(
  MR_all_age1$ci.ub, 
  MR_all_age$ci.ub[2], 
  MR_amb_age1$ci.ub, 
  MR_amb_age$ci.ub[2], 
  MR_MID_age1$ci.ub, 
  MR_MID_age$ci.ub[2], 
  MR_maxabsES_age1$ci.ub, 
  MR_maxabsES_age$ci.ub[2], 
  MR_maxmeanES_age1$ci.ub, 
  MR_maxmeanES_age$ci.ub[2], 
  MR_maxES_age1$ci.ub, 
  MR_maxES_age$ci.ub[2]), 3) 

t14$R2[pos] <- c(
  table(dat$age), 
  rep(NA, 1),
  table(dat_amb$age), 
  rep(NA, 1),
  table(dat_MID_ES$age), 
  rep(NA, 1),
  table(dat_max_absES$age), 
  rep(NA, 1),
  table(dat_max_meanES$age), 
  rep(NA, 1),
  table(dat_max_ES$age), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t14$`R2`[posR2] <- round(c(R2(MR_all_age)[1], R2(MR_amb_age)[1], R2(MR_MID_age)[1], R2(MR_maxabsES_age)[1], R2(MR_maxmeanES_age)[1], R2(MR_maxES_age)[1]),3)

write.csv(t14, file = "./Tables/Table_S14_MR_Age.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S15      
MR - univariate AffectManipCat model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ AffectManipCat}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_AffectManipCat1.Rdata")
load("./Rdata/MR_all_AffectManipCat.Rdata") 
load("./Rdata/MR_amb_AffectManipCat1.Rdata")
load("./Rdata/MR_amb_AffectManipCat.Rdata")
load("./Rdata/MR_MID_AffectManipCat1.Rdata")
load("./Rdata/MR_MID_AffectManipCat.Rdata")
load("./Rdata/MR_maxabsES_AffectManipCat1.Rdata")
load("./Rdata/MR_maxabsES_AffectManipCat.Rdata")
load("./Rdata/MR_maxmeanES_AffectManipCat1.Rdata")
load("./Rdata/MR_maxmeanES_AffectManipCat.Rdata")
load("./Rdata/MR_maxES_AffectManipCat1.Rdata")
load("./Rdata/MR_maxES_AffectManipCat.Rdata")

# loading table
t15 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 12)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t15$`Mean`[pos] <- round(c(
  MR_all_AffectManipCat1$b, 
  MR_all_AffectManipCat$b[2], 
  MR_amb_AffectManipCat1$b, 
  MR_amb_AffectManipCat$b[2], 
  MR_MID_AffectManipCat1$b, 
  MR_MID_AffectManipCat$b[2], 
  MR_maxabsES_AffectManipCat1$b, 
  MR_maxabsES_AffectManipCat$b[2], 
  MR_maxmeanES_AffectManipCat1$b, 
  MR_maxmeanES_AffectManipCat$b[2], 
  MR_maxES_AffectManipCat1$b, 
  MR_maxES_AffectManipCat$b[2]), 3) 
  
t15$CI.lb[pos] <- round(c(
  MR_all_AffectManipCat1$ci.lb, 
  MR_all_AffectManipCat$ci.lb[2], 
  MR_amb_AffectManipCat1$ci.lb, 
  MR_amb_AffectManipCat$ci.lb[2], 
  MR_MID_AffectManipCat1$ci.lb, 
  MR_MID_AffectManipCat$ci.lb[2], 
  MR_maxabsES_AffectManipCat1$ci.lb, 
  MR_maxabsES_AffectManipCat$ci.lb[2], 
  MR_maxmeanES_AffectManipCat1$ci.lb, 
  MR_maxmeanES_AffectManipCat$ci.lb[2], 
  MR_maxES_AffectManipCat1$ci.lb, 
  MR_maxES_AffectManipCat$ci.lb[2]), 3) 

t15$CI.ub[pos] <- round(c(
  MR_all_AffectManipCat1$ci.ub, 
  MR_all_AffectManipCat$ci.ub[2], 
  MR_amb_AffectManipCat1$ci.ub, 
  MR_amb_AffectManipCat$ci.ub[2], 
  MR_MID_AffectManipCat1$ci.ub, 
  MR_MID_AffectManipCat$ci.ub[2], 
  MR_maxabsES_AffectManipCat1$ci.ub, 
  MR_maxabsES_AffectManipCat$ci.ub[2], 
  MR_maxmeanES_AffectManipCat1$ci.ub, 
  MR_maxmeanES_AffectManipCat$ci.ub[2], 
  MR_maxES_AffectManipCat1$ci.ub, 
  MR_maxES_AffectManipCat$ci.ub[2]), 3) 

t15$R2[pos] <- c(
  table(dat$AffectManipCat), 
  rep(NA, 1),
  table(dat_amb$AffectManipCat), 
  rep(NA, 1),
  table(dat_MID_ES$AffectManipCat), 
  rep(NA, 1),
  table(dat_max_absES$AffectManipCat), 
  rep(NA, 1),
  table(dat_max_meanES$AffectManipCat), 
  rep(NA, 1),
  table(dat_max_ES$AffectManipCat), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t15$`R2`[posR2] <- round(c(R2(MR_all_AffectManipCat)[1], R2(MR_amb_AffectManipCat)[1], R2(MR_MID_AffectManipCat)[1], R2(MR_maxabsES_AffectManipCat)[1], R2(MR_maxmeanES_AffectManipCat)[1], R2(MR_maxES_AffectManipCat)[1]),3)

write.csv(t15, file = "./Tables/Table_S15_MR_AffectManipCat.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S16      
MR - univariate AffectManipTiming model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ AffectManipTiming}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_AffectManipTiming1.Rdata")
load("./Rdata/MR_all_AffectManipTiming.Rdata") 
load("./Rdata/MR_amb_AffectManipTiming1.Rdata")
load("./Rdata/MR_amb_AffectManipTiming.Rdata")
load("./Rdata/MR_MID_AffectManipTiming1.Rdata")
load("./Rdata/MR_MID_AffectManipTiming.Rdata")
load("./Rdata/MR_maxabsES_AffectManipTiming1.Rdata")
load("./Rdata/MR_maxabsES_AffectManipTiming.Rdata")
load("./Rdata/MR_maxmeanES_AffectManipTiming1.Rdata")
load("./Rdata/MR_maxmeanES_AffectManipTiming.Rdata")
load("./Rdata/MR_maxES_AffectManipTiming1.Rdata")
load("./Rdata/MR_maxES_AffectManipTiming.Rdata")

# loading table
t16 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 13)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t16$`Mean`[pos] <- round(c(
  MR_all_AffectManipTiming1$b, 
  MR_all_AffectManipTiming$b[2], 
  MR_amb_AffectManipTiming1$b, 
  MR_amb_AffectManipTiming$b[2], 
  MR_MID_AffectManipTiming1$b, 
  MR_MID_AffectManipTiming$b[2], 
  MR_maxabsES_AffectManipTiming1$b, 
  MR_maxabsES_AffectManipTiming$b[2], 
  MR_maxmeanES_AffectManipTiming1$b, 
  MR_maxmeanES_AffectManipTiming$b[2], 
  MR_maxES_AffectManipTiming1$b, 
  MR_maxES_AffectManipTiming$b[2]), 3) 
  
t16$CI.lb[pos] <- round(c(
  MR_all_AffectManipTiming1$ci.lb, 
  MR_all_AffectManipTiming$ci.lb[2], 
  MR_amb_AffectManipTiming1$ci.lb, 
  MR_amb_AffectManipTiming$ci.lb[2], 
  MR_MID_AffectManipTiming1$ci.lb, 
  MR_MID_AffectManipTiming$ci.lb[2], 
  MR_maxabsES_AffectManipTiming1$ci.lb, 
  MR_maxabsES_AffectManipTiming$ci.lb[2], 
  MR_maxmeanES_AffectManipTiming1$ci.lb, 
  MR_maxmeanES_AffectManipTiming$ci.lb[2], 
  MR_maxES_AffectManipTiming1$ci.lb, 
  MR_maxES_AffectManipTiming$ci.lb[2]), 3) 

t16$CI.ub[pos] <- round(c(
  MR_all_AffectManipTiming1$ci.ub, 
  MR_all_AffectManipTiming$ci.ub[2], 
  MR_amb_AffectManipTiming1$ci.ub, 
  MR_amb_AffectManipTiming$ci.ub[2], 
  MR_MID_AffectManipTiming1$ci.ub, 
  MR_MID_AffectManipTiming$ci.ub[2], 
  MR_maxabsES_AffectManipTiming1$ci.ub, 
  MR_maxabsES_AffectManipTiming$ci.ub[2], 
  MR_maxmeanES_AffectManipTiming1$ci.ub, 
  MR_maxmeanES_AffectManipTiming$ci.ub[2], 
  MR_maxES_AffectManipTiming1$ci.ub, 
  MR_maxES_AffectManipTiming$ci.ub[2]), 3) 

t16$R2[pos] <- c(
  table(dat$AffectManipTiming), 
  rep(NA, 1),
  table(dat_amb$AffectManipTiming), 
  rep(NA, 1),
  table(dat_MID_ES$AffectManipTiming), 
  rep(NA, 1),
  table(dat_max_absES$AffectManipTiming), 
  rep(NA, 1),
  table(dat_max_meanES$AffectManipTiming), 
  rep(NA, 1),
  table(dat_max_ES$AffectManipTiming), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t16$`R2`[posR2] <- round(c(R2(MR_all_AffectManipTiming)[1], R2(MR_amb_AffectManipTiming)[1], R2(MR_MID_AffectManipTiming)[1], R2(MR_maxabsES_AffectManipTiming)[1], R2(MR_maxmeanES_AffectManipTiming)[1], R2(MR_maxES_AffectManipTiming)[1]),3)

write.csv(t16, file = "./Tables/Table_S16_MR_AffectManipTiming.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S17      
MR - univariate ComparisonCategory model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ ComparisonCategory}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_ComparisonCategory1.Rdata")
load("./Rdata/MR_all_ComparisonCategory.Rdata") 
load("./Rdata/MR_all_ComparisonCategory2.Rdata") 
load("./Rdata/MR_amb_ComparisonCategory1.Rdata")
load("./Rdata/MR_amb_ComparisonCategory.Rdata")
load("./Rdata/MR_amb_ComparisonCategory2.Rdata")
load("./Rdata/MR_MID_ComparisonCategory1.Rdata")
load("./Rdata/MR_MID_ComparisonCategory.Rdata")
load("./Rdata/MR_MID_ComparisonCategory2.Rdata")
load("./Rdata/MR_maxabsES_ComparisonCategory1.Rdata")
load("./Rdata/MR_maxabsES_ComparisonCategory.Rdata")
load("./Rdata/MR_maxabsES_ComparisonCategory2.Rdata")
load("./Rdata/MR_maxmeanES_ComparisonCategory1.Rdata")
load("./Rdata/MR_maxmeanES_ComparisonCategory.Rdata")
load("./Rdata/MR_maxmeanES_ComparisonCategory2.Rdata")
load("./Rdata/MR_maxES_ComparisonCategory1.Rdata")
load("./Rdata/MR_maxES_ComparisonCategory.Rdata")
load("./Rdata/MR_maxES_ComparisonCategory2.Rdata")

# loading table
t17 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 14)

# saving row position for species estimates
pos <- c(2:7, 9:14, 16:21, 23:28, 30:35, 37:42)

t17$`Mean`[pos] <- round(c(
  MR_all_ComparisonCategory1$b, 
  MR_all_ComparisonCategory$b[2:3], 
  MR_all_ComparisonCategory2$b[3], 
  MR_amb_ComparisonCategory1$b, 
  MR_amb_ComparisonCategory$b[2:3], 
  MR_amb_ComparisonCategory2$b[3], 
  MR_MID_ComparisonCategory1$b, 
  MR_MID_ComparisonCategory$b[2:3], 
  MR_MID_ComparisonCategory2$b[3], 
  MR_maxabsES_ComparisonCategory1$b, 
  MR_maxabsES_ComparisonCategory$b[2:3], 
  MR_maxabsES_ComparisonCategory2$b[3], 
  MR_maxmeanES_ComparisonCategory1$b, 
  MR_maxmeanES_ComparisonCategory$b[2:3], 
  MR_maxmeanES_ComparisonCategory2$b[3], 
  MR_maxES_ComparisonCategory1$b, 
  MR_maxES_ComparisonCategory$b[2:3], 
  MR_maxES_ComparisonCategory2$b[3]), 3) 
  
t17$CI.lb[pos] <- round(c(
  MR_all_ComparisonCategory1$ci.lb, 
  MR_all_ComparisonCategory$ci.lb[2:3], 
  MR_all_ComparisonCategory2$ci.lb[3], 
  MR_amb_ComparisonCategory1$ci.lb, 
  MR_amb_ComparisonCategory$ci.lb[2:3], 
  MR_amb_ComparisonCategory2$ci.lb[3], 
  MR_MID_ComparisonCategory1$ci.lb, 
  MR_MID_ComparisonCategory$ci.lb[2:3], 
  MR_MID_ComparisonCategory2$ci.lb[3], 
  MR_maxabsES_ComparisonCategory1$ci.lb, 
  MR_maxabsES_ComparisonCategory$ci.lb[2:3], 
  MR_maxabsES_ComparisonCategory2$ci.lb[3], 
  MR_maxmeanES_ComparisonCategory1$ci.lb, 
  MR_maxmeanES_ComparisonCategory$ci.lb[2:3], 
  MR_maxmeanES_ComparisonCategory2$ci.lb[3], 
  MR_maxES_ComparisonCategory1$ci.lb, 
  MR_maxES_ComparisonCategory$ci.lb[2:3], 
  MR_maxES_ComparisonCategory2$ci.lb[3]), 3) 

t17$CI.ub[pos] <- round(c(
  MR_all_ComparisonCategory1$ci.ub, 
  MR_all_ComparisonCategory$ci.ub[2:3], 
  MR_all_ComparisonCategory2$ci.ub[3], 
  MR_amb_ComparisonCategory1$ci.ub, 
  MR_amb_ComparisonCategory$ci.ub[2:3], 
  MR_amb_ComparisonCategory2$ci.ub[3], 
  MR_MID_ComparisonCategory1$ci.ub, 
  MR_MID_ComparisonCategory$ci.ub[2:3], 
  MR_MID_ComparisonCategory2$ci.ub[3], 
  MR_maxabsES_ComparisonCategory1$ci.ub, 
  MR_maxabsES_ComparisonCategory$ci.ub[2:3], 
  MR_maxabsES_ComparisonCategory2$ci.ub[3], 
  MR_maxmeanES_ComparisonCategory1$ci.ub, 
  MR_maxmeanES_ComparisonCategory$ci.ub[2:3], 
  MR_maxmeanES_ComparisonCategory2$ci.ub[3], 
  MR_maxES_ComparisonCategory1$ci.ub, 
  MR_maxES_ComparisonCategory$ci.ub[2:3], 
  MR_maxES_ComparisonCategory2$ci.ub[3]), 3) 

t17$R2[pos] <- c(
  table(dat$ComparisonCategory), 
  rep(NA, 3),
  table(dat_amb$ComparisonCategory), 
  rep(NA, 3),
  table(dat_MID_ES$ComparisonCategory), 
  rep(NA, 3),
  table(dat_max_absES$ComparisonCategory), 
  rep(NA, 3),
  table(dat_max_meanES$ComparisonCategory), 
  rep(NA, 3),
  table(dat_max_ES$ComparisonCategory), 
  rep(NA, 3)) 

# adding model R2 estimates to the table
posR2 <- c(1, 8, 15, 22, 29, 36) #row positions
t17$`R2`[posR2] <- round(c(R2(MR_all_ComparisonCategory)[1], R2(MR_amb_ComparisonCategory)[1], R2(MR_MID_ComparisonCategory)[1], R2(MR_maxabsES_ComparisonCategory)[1], R2(MR_maxmeanES_ComparisonCategory)[1], R2(MR_maxES_ComparisonCategory)[1]),3)

write.csv(t17, file = "./Tables/Table_S17_MR_ComparisonCategory.csv", row.names = F, na = "")    
```



###################################################################    
#### Table S18      
MR - univariate WithinBetween model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ WithinBetween}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_WithinBetween1.Rdata")
load("./Rdata/MR_all_WithinBetween.Rdata") 
load("./Rdata/MR_amb_WithinBetween1.Rdata")
load("./Rdata/MR_amb_WithinBetween.Rdata")
load("./Rdata/MR_MID_WithinBetween1.Rdata")
load("./Rdata/MR_MID_WithinBetween.Rdata")
load("./Rdata/MR_maxabsES_WithinBetween1.Rdata")
load("./Rdata/MR_maxabsES_WithinBetween.Rdata")
load("./Rdata/MR_maxmeanES_WithinBetween1.Rdata")
load("./Rdata/MR_maxmeanES_WithinBetween.Rdata")
load("./Rdata/MR_maxES_WithinBetween1.Rdata")
load("./Rdata/MR_maxES_WithinBetween.Rdata")

# loading table
t18 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 15)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t18$`Mean`[pos] <- round(c(
  MR_all_WithinBetween1$b, 
  MR_all_WithinBetween$b[2], 
  MR_amb_WithinBetween1$b, 
  MR_amb_WithinBetween$b[2], 
  MR_MID_WithinBetween1$b, 
  MR_MID_WithinBetween$b[2], 
  MR_maxabsES_WithinBetween1$b, 
  MR_maxabsES_WithinBetween$b[2], 
  MR_maxmeanES_WithinBetween1$b, 
  MR_maxmeanES_WithinBetween$b[2], 
  MR_maxES_WithinBetween1$b, 
  MR_maxES_WithinBetween$b[2]), 3) 
  
t18$CI.lb[pos] <- round(c(
  MR_all_WithinBetween1$ci.lb, 
  MR_all_WithinBetween$ci.lb[2], 
  MR_amb_WithinBetween1$ci.lb, 
  MR_amb_WithinBetween$ci.lb[2], 
  MR_MID_WithinBetween1$ci.lb, 
  MR_MID_WithinBetween$ci.lb[2], 
  MR_maxabsES_WithinBetween1$ci.lb, 
  MR_maxabsES_WithinBetween$ci.lb[2], 
  MR_maxmeanES_WithinBetween1$ci.lb, 
  MR_maxmeanES_WithinBetween$ci.lb[2], 
  MR_maxES_WithinBetween1$ci.lb, 
  MR_maxES_WithinBetween$ci.lb[2]), 3) 

t18$CI.ub[pos] <- round(c(
  MR_all_WithinBetween1$ci.ub, 
  MR_all_WithinBetween$ci.ub[2], 
  MR_amb_WithinBetween1$ci.ub, 
  MR_amb_WithinBetween$ci.ub[2], 
  MR_MID_WithinBetween1$ci.ub, 
  MR_MID_WithinBetween$ci.ub[2], 
  MR_maxabsES_WithinBetween1$ci.ub, 
  MR_maxabsES_WithinBetween$ci.ub[2], 
  MR_maxmeanES_WithinBetween1$ci.ub, 
  MR_maxmeanES_WithinBetween$ci.ub[2], 
  MR_maxES_WithinBetween1$ci.ub, 
  MR_maxES_WithinBetween$ci.ub[2]), 3) 

t18$R2[pos] <- c(
  table(dat$WithinBetween), 
  rep(NA, 1),
  table(dat_amb$WithinBetween), 
  rep(NA, 1),
  table(dat_MID_ES$WithinBetween), 
  rep(NA, 1),
  table(dat_max_absES$WithinBetween), 
  rep(NA, 1),
  table(dat_max_meanES$WithinBetween), 
  rep(NA, 1),
  table(dat_max_ES$WithinBetween), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t18$`R2`[posR2] <- round(c(R2(MR_all_WithinBetween)[1], R2(MR_amb_WithinBetween)[1], R2(MR_MID_WithinBetween)[1], R2(MR_maxabsES_WithinBetween)[1], R2(MR_maxmeanES_WithinBetween)[1], R2(MR_maxES_WithinBetween)[1]),3)

write.csv(t18, file = "./Tables/Table_S18_MR_WithinBetween.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S19      
MR - univariate FoodDeprived model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ FoodDeprived}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_FoodDeprived1.Rdata")
load("./Rdata/MR_all_FoodDeprived.Rdata") 
load("./Rdata/MR_amb_FoodDeprived1.Rdata")
load("./Rdata/MR_amb_FoodDeprived.Rdata")
load("./Rdata/MR_MID_FoodDeprived1.Rdata")
load("./Rdata/MR_MID_FoodDeprived.Rdata")
load("./Rdata/MR_maxabsES_FoodDeprived1.Rdata")
load("./Rdata/MR_maxabsES_FoodDeprived.Rdata")
load("./Rdata/MR_maxmeanES_FoodDeprived1.Rdata")
load("./Rdata/MR_maxmeanES_FoodDeprived.Rdata")
load("./Rdata/MR_maxES_FoodDeprived1.Rdata")
load("./Rdata/MR_maxES_FoodDeprived.Rdata")

# loading table
t19 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 16)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t19$`Mean`[pos] <- round(c(
  MR_all_FoodDeprived1$b, 
  MR_all_FoodDeprived$b[2], 
  MR_amb_FoodDeprived1$b, 
  MR_amb_FoodDeprived$b[2], 
  MR_MID_FoodDeprived1$b, 
  MR_MID_FoodDeprived$b[2], 
  MR_maxabsES_FoodDeprived1$b, 
  MR_maxabsES_FoodDeprived$b[2], 
  MR_maxmeanES_FoodDeprived1$b, 
  MR_maxmeanES_FoodDeprived$b[2], 
  MR_maxES_FoodDeprived1$b, 
  MR_maxES_FoodDeprived$b[2]), 3) 
  
t19$CI.lb[pos] <- round(c(
  MR_all_FoodDeprived1$ci.lb, 
  MR_all_FoodDeprived$ci.lb[2], 
  MR_amb_FoodDeprived1$ci.lb, 
  MR_amb_FoodDeprived$ci.lb[2], 
  MR_MID_FoodDeprived1$ci.lb, 
  MR_MID_FoodDeprived$ci.lb[2], 
  MR_maxabsES_FoodDeprived1$ci.lb, 
  MR_maxabsES_FoodDeprived$ci.lb[2], 
  MR_maxmeanES_FoodDeprived1$ci.lb, 
  MR_maxmeanES_FoodDeprived$ci.lb[2], 
  MR_maxES_FoodDeprived1$ci.lb, 
  MR_maxES_FoodDeprived$ci.lb[2]), 3) 

t19$CI.ub[pos] <- round(c(
  MR_all_FoodDeprived1$ci.ub, 
  MR_all_FoodDeprived$ci.ub[2], 
  MR_amb_FoodDeprived1$ci.ub, 
  MR_amb_FoodDeprived$ci.ub[2], 
  MR_MID_FoodDeprived1$ci.ub, 
  MR_MID_FoodDeprived$ci.ub[2], 
  MR_maxabsES_FoodDeprived1$ci.ub, 
  MR_maxabsES_FoodDeprived$ci.ub[2], 
  MR_maxmeanES_FoodDeprived1$ci.ub, 
  MR_maxmeanES_FoodDeprived$ci.ub[2], 
  MR_maxES_FoodDeprived1$ci.ub, 
  MR_maxES_FoodDeprived$ci.ub[2]), 3) 

t19$R2[pos] <- c(
  table(dat$FoodDeprived), 
  rep(NA, 1),
  table(dat_amb$FoodDeprived), 
  rep(NA, 1),
  table(dat_MID_ES$FoodDeprived), 
  rep(NA, 1),
  table(dat_max_absES$FoodDeprived), 
  rep(NA, 1),
  table(dat_max_meanES$FoodDeprived), 
  rep(NA, 1),
  table(dat_max_ES$FoodDeprived), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t19$`R2`[posR2] <- round(c(R2(MR_all_FoodDeprived)[1], R2(MR_amb_FoodDeprived)[1], R2(MR_MID_FoodDeprived)[1], R2(MR_maxabsES_FoodDeprived)[1], R2(MR_maxmeanES_FoodDeprived)[1], R2(MR_maxES_FoodDeprived)[1]),3)

write.csv(t19, file = "./Tables/Table_S19_MR_FoodDeprived.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S20      
MR - univariate MeasureType model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ MeasureType}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_MeasureType1.Rdata")
load("./Rdata/MR_all_MeasureType.Rdata") 
load("./Rdata/MR_amb_MeasureType1.Rdata")
load("./Rdata/MR_amb_MeasureType.Rdata")
load("./Rdata/MR_MID_MeasureType1.Rdata")
load("./Rdata/MR_MID_MeasureType.Rdata")
load("./Rdata/MR_maxabsES_MeasureType1.Rdata")
load("./Rdata/MR_maxabsES_MeasureType.Rdata")
load("./Rdata/MR_maxmeanES_MeasureType1.Rdata")
load("./Rdata/MR_maxmeanES_MeasureType.Rdata")
load("./Rdata/MR_maxES_MeasureType1.Rdata")
load("./Rdata/MR_maxES_MeasureType.Rdata")

# loading table
t20 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 17)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t20$`Mean`[pos] <- round(c(
  MR_all_MeasureType1$b, 
  MR_all_MeasureType$b[2], 
  MR_amb_MeasureType1$b, 
  MR_amb_MeasureType$b[2], 
  MR_MID_MeasureType1$b, 
  MR_MID_MeasureType$b[2], 
  MR_maxabsES_MeasureType1$b, 
  MR_maxabsES_MeasureType$b[2], 
  MR_maxmeanES_MeasureType1$b, 
  MR_maxmeanES_MeasureType$b[2], 
  MR_maxES_MeasureType1$b, 
  MR_maxES_MeasureType$b[2]), 3) 
  
t20$CI.lb[pos] <- round(c(
  MR_all_MeasureType1$ci.lb, 
  MR_all_MeasureType$ci.lb[2], 
  MR_amb_MeasureType1$ci.lb, 
  MR_amb_MeasureType$ci.lb[2], 
  MR_MID_MeasureType1$ci.lb, 
  MR_MID_MeasureType$ci.lb[2], 
  MR_maxabsES_MeasureType1$ci.lb, 
  MR_maxabsES_MeasureType$ci.lb[2], 
  MR_maxmeanES_MeasureType1$ci.lb, 
  MR_maxmeanES_MeasureType$ci.lb[2], 
  MR_maxES_MeasureType1$ci.lb, 
  MR_maxES_MeasureType$ci.lb[2]), 3) 

t20$CI.ub[pos] <- round(c(
  MR_all_MeasureType1$ci.ub, 
  MR_all_MeasureType$ci.ub[2], 
  MR_amb_MeasureType1$ci.ub, 
  MR_amb_MeasureType$ci.ub[2], 
  MR_MID_MeasureType1$ci.ub, 
  MR_MID_MeasureType$ci.ub[2], 
  MR_maxabsES_MeasureType1$ci.ub, 
  MR_maxabsES_MeasureType$ci.ub[2], 
  MR_maxmeanES_MeasureType1$ci.ub, 
  MR_maxmeanES_MeasureType$ci.ub[2], 
  MR_maxES_MeasureType1$ci.ub, 
  MR_maxES_MeasureType$ci.ub[2]), 3) 

t20$R2[pos] <- c(
  table(dat$MeasureType), 
  rep(NA, 1),
  table(dat_amb$MeasureType), 
  rep(NA, 1),
  table(dat_MID_ES$MeasureType), 
  rep(NA, 1),
  table(dat_max_absES$MeasureType), 
  rep(NA, 1),
  table(dat_max_meanES$MeasureType), 
  rep(NA, 1),
  table(dat_max_ES$MeasureType), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t20$`R2`[posR2] <- round(c(R2(MR_all_MeasureType)[1], R2(MR_amb_MeasureType)[1], R2(MR_MID_MeasureType)[1], R2(MR_maxabsES_MeasureType)[1], R2(MR_maxmeanES_MeasureType)[1], R2(MR_maxES_MeasureType)[1]),3)

write.csv(t20, file = "./Tables/Table_S20_MR_MeasureType.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S21      
MR - univariate Automated model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Automated}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_Automated1.Rdata")
load("./Rdata/MR_all_Automated.Rdata") 
load("./Rdata/MR_amb_Automated1.Rdata")
load("./Rdata/MR_amb_Automated.Rdata")
load("./Rdata/MR_MID_Automated1.Rdata")
load("./Rdata/MR_MID_Automated.Rdata")
load("./Rdata/MR_maxabsES_Automated1.Rdata")
load("./Rdata/MR_maxabsES_Automated.Rdata")
load("./Rdata/MR_maxmeanES_Automated1.Rdata")
load("./Rdata/MR_maxmeanES_Automated.Rdata")
load("./Rdata/MR_maxES_Automated1.Rdata")
load("./Rdata/MR_maxES_Automated.Rdata")

# loading table
t21 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 18)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t21$`Mean`[pos] <- round(c(
  MR_all_Automated1$b, 
  MR_all_Automated$b[2], 
  MR_amb_Automated1$b, 
  MR_amb_Automated$b[2], 
  MR_MID_Automated1$b, 
  MR_MID_Automated$b[2], 
  MR_maxabsES_Automated1$b, 
  MR_maxabsES_Automated$b[2], 
  MR_maxmeanES_Automated1$b, 
  MR_maxmeanES_Automated$b[2], 
  MR_maxES_Automated1$b, 
  MR_maxES_Automated$b[2]), 3) 
  
t21$CI.lb[pos] <- round(c(
  MR_all_Automated1$ci.lb, 
  MR_all_Automated$ci.lb[2], 
  MR_amb_Automated1$ci.lb, 
  MR_amb_Automated$ci.lb[2], 
  MR_MID_Automated1$ci.lb, 
  MR_MID_Automated$ci.lb[2], 
  MR_maxabsES_Automated1$ci.lb, 
  MR_maxabsES_Automated$ci.lb[2], 
  MR_maxmeanES_Automated1$ci.lb, 
  MR_maxmeanES_Automated$ci.lb[2], 
  MR_maxES_Automated1$ci.lb, 
  MR_maxES_Automated$ci.lb[2]), 3) 

t21$CI.ub[pos] <- round(c(
  MR_all_Automated1$ci.ub, 
  MR_all_Automated$ci.ub[2], 
  MR_amb_Automated1$ci.ub, 
  MR_amb_Automated$ci.ub[2], 
  MR_MID_Automated1$ci.ub, 
  MR_MID_Automated$ci.ub[2], 
  MR_maxabsES_Automated1$ci.ub, 
  MR_maxabsES_Automated$ci.ub[2], 
  MR_maxmeanES_Automated1$ci.ub, 
  MR_maxmeanES_Automated$ci.ub[2], 
  MR_maxES_Automated1$ci.ub, 
  MR_maxES_Automated$ci.ub[2]), 3) 

t21$R2[pos] <- c(
  table(dat$Automated), 
  rep(NA, 1),
  table(dat_amb$Automated), 
  rep(NA, 1),
  table(dat_MID_ES$Automated), 
  rep(NA, 1),
  table(dat_max_absES$Automated), 
  rep(NA, 1),
  table(dat_max_meanES$Automated), 
  rep(NA, 1),
  table(dat_max_ES$Automated), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t21$`R2`[posR2] <- round(c(R2(MR_all_Automated)[1], R2(MR_amb_Automated)[1], R2(MR_MID_Automated)[1], R2(MR_maxabsES_Automated)[1], R2(MR_maxmeanES_Automated)[1], R2(MR_maxES_Automated)[1]),3)

write.csv(t21, file = "./Tables/Table_S21_MR_Automated.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S22      
MR - univariate Blind model - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Blind}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_Blind1.Rdata")
load("./Rdata/MR_all_Blind.Rdata") 
load("./Rdata/MR_amb_Blind1.Rdata")
load("./Rdata/MR_amb_Blind.Rdata")
load("./Rdata/MR_MID_Blind1.Rdata")
load("./Rdata/MR_MID_Blind.Rdata")
load("./Rdata/MR_maxabsES_Blind1.Rdata")
load("./Rdata/MR_maxabsES_Blind.Rdata")
load("./Rdata/MR_maxmeanES_Blind1.Rdata")
load("./Rdata/MR_maxmeanES_Blind.Rdata")
load("./Rdata/MR_maxES_Blind1.Rdata")
load("./Rdata/MR_maxES_Blind.Rdata")

# loading table
t22 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 19)

# saving row position for species estimates
pos <- c(2:4, 6:8, 10:12, 14:16, 18:20, 22:24)

t22$`Mean`[pos] <- round(c(
  MR_all_Blind1$b, 
  MR_all_Blind$b[2], 
  MR_amb_Blind1$b, 
  MR_amb_Blind$b[2], 
  MR_MID_Blind1$b, 
  MR_MID_Blind$b[2], 
  MR_maxabsES_Blind1$b, 
  MR_maxabsES_Blind$b[2], 
  MR_maxmeanES_Blind1$b, 
  MR_maxmeanES_Blind$b[2], 
  MR_maxES_Blind1$b, 
  MR_maxES_Blind$b[2]), 3) 
  
t22$CI.lb[pos] <- round(c(
  MR_all_Blind1$ci.lb, 
  MR_all_Blind$ci.lb[2], 
  MR_amb_Blind1$ci.lb, 
  MR_amb_Blind$ci.lb[2], 
  MR_MID_Blind1$ci.lb, 
  MR_MID_Blind$ci.lb[2], 
  MR_maxabsES_Blind1$ci.lb, 
  MR_maxabsES_Blind$ci.lb[2], 
  MR_maxmeanES_Blind1$ci.lb, 
  MR_maxmeanES_Blind$ci.lb[2], 
  MR_maxES_Blind1$ci.lb, 
  MR_maxES_Blind$ci.lb[2]), 3) 

t22$CI.ub[pos] <- round(c(
  MR_all_Blind1$ci.ub, 
  MR_all_Blind$ci.ub[2], 
  MR_amb_Blind1$ci.ub, 
  MR_amb_Blind$ci.ub[2], 
  MR_MID_Blind1$ci.ub, 
  MR_MID_Blind$ci.ub[2], 
  MR_maxabsES_Blind1$ci.ub, 
  MR_maxabsES_Blind$ci.ub[2], 
  MR_maxmeanES_Blind1$ci.ub, 
  MR_maxmeanES_Blind$ci.ub[2], 
  MR_maxES_Blind1$ci.ub, 
  MR_maxES_Blind$ci.ub[2]), 3) 

t22$R2[pos] <- c(
  table(dat$Blind), 
  rep(NA, 1),
  table(dat_amb$Blind), 
  rep(NA, 1),
  table(dat_MID_ES$Blind), 
  rep(NA, 1),
  table(dat_max_absES$Blind), 
  rep(NA, 1),
  table(dat_max_meanES$Blind), 
  rep(NA, 1),
  table(dat_max_ES$Blind), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 5, 9, 13, 17, 21) #row positions
t22$`R2`[posR2] <- round(c(R2(MR_all_Blind)[1], R2(MR_amb_Blind)[1], R2(MR_MID_Blind)[1], R2(MR_maxabsES_Blind)[1], R2(MR_maxmeanES_Blind)[1], R2(MR_maxES_Blind)[1]),3)

write.csv(t22, file = "./Tables/Table_S22_MR_Blind.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S23      
MR - multivariate (full) model - with VCV and phylo - all data and 5 data subsets  

```{r Table MR univ Year}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_Full_0.Rdata") 
load("./Rdata/MR_Full_0_ambES.Rdata")
load("./Rdata/MR_Full_0_MID_ES.Rdata")
load("./Rdata/MR_Full_0_max_absES.Rdata")
load("./Rdata/MR_Full_0_max_meanES.Rdata")
load("./Rdata/MR_Full_0_max_ES.Rdata")

# loading table
t23 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 20)

# saving row position for species estimates
pos <- c(2:11, 13:22, 24:33, 35:44, 46:55, 57:66)

t23$`Mean`[pos] <- round(c(
  MR_Full_0$b, 
  MR_Full_0_ambES$b, 
  MR_Full_0_MID_ES$b, 
  MR_Full_0_max_absES$b, 
  MR_Full_0_max_meanES$b, 
  MR_Full_0_max_ES$b), 3) 
  
t23$CI.lb[pos] <- round(c(
  MR_Full_0$ci.lb, 
  MR_Full_0_ambES$ci.lb, 
  MR_Full_0_MID_ES$ci.lb, 
  MR_Full_0_max_absES$ci.lb, 
  MR_Full_0_max_meanES$ci.lb, 
  MR_Full_0_max_ES$ci.lb), 3) 

t23$CI.ub[pos] <- round(c(
  MR_Full_0$ci.ub, 
  MR_Full_0_ambES$ci.ub, 
  MR_Full_0_MID_ES$ci.ub, 
  MR_Full_0_max_absES$ci.ub, 
  MR_Full_0_max_meanES$ci.ub, 
  MR_Full_0_max_ES$ci.ub), 3) 

# adding model R2 estimates to the table
posR2 <- c(1, 12, 23, 34, 45, 56) #row positions
t23$`R2`[posR2] <- round(c(R2(MR_Full_0)[1],
                           R2(MR_Full_0_ambES)[1], 
                           R2(MR_Full_0_MID_ES)[1], 
                           R2(MR_Full_0_max_absES)[1], 
                           R2(MR_Full_0_max_meanES)[1], 
                           R2(MR_Full_0_max_ES)[1]),3)

write.csv(t23, file = "./Tables/Table_S23_MR_full_model.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S24      
model selection results - all data     

```{r Table model selection}
#list.files("./Rdata/")
# loading data
load(file = "./Rdata/meta_AICc_seltable.Rdata")
sel.table

# loading table
t24 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 21)

# saving row position for species estimates
pos <- c(1:nrow(sel.table))

t24$`Model`[pos] <- sel.table$Model 
t24$`K`[pos] <- sel.table$K 
t24$`logLik`[pos] <- sel.table$logLik 
t24$`AICc`[pos] <- sel.table$AICc 
t24$`delta`[pos] <- sel.table$delta 
t24$`weight`[pos] <- sel.table$weight 

write.csv(t24, file = "./Tables/Table_S24_model_selection.csv", row.names = F, na = "")    
```

###################################################################    
#### Table S25      
MR - univariate Year model (time-lag publication bias) - with VCV and phylo - all data and 5 data subsets    

```{r Table MR univ Year}
#list.files("./Rdata/")
# loading data
load("./Rdata/MR_all_Year.Rdata") 
load("./Rdata/MR_amb_Year.Rdata")
load("./Rdata/MR_MID_Year.Rdata")
load("./Rdata/MR_maxabsES_Year.Rdata")
load("./Rdata/MR_maxmeanES_Year.Rdata")
load("./Rdata/MR_maxES_Year.Rdata")

# loading table
t25 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 22)

# saving row position for species estimates
pos <- c(2:3, 5:6, 8:9, 11:12, 14:15, 17:18)

t25$`Mean`[pos] <- round(c(
  MR_all_Year$b, 
  MR_amb_Year$b, 
  MR_MID_Year$b, 
  MR_maxabsES_Year$b, 
  MR_maxmeanES_Year$b, 
  MR_maxES_Year$b), 3) 
  
t25$CI.lb[pos] <- round(c(
  MR_all_Year$ci.lb, 
  MR_amb_Year$ci.lb, 
  MR_MID_Year$ci.lb, 
  MR_maxabsES_Year$ci.lb, 
  MR_maxmeanES_Year$ci.lb, 
  MR_maxES_Year$ci.lb), 3) 

t25$CI.ub[pos] <- round(c(
  MR_all_Year$ci.ub, 
  MR_amb_Year$ci.ub, 
  MR_MID_Year$ci.ub, 
  MR_maxabsES_Year$ci.ub, 
  MR_maxmeanES_Year$ci.ub, 
  MR_maxES_Year$ci.ub), 3) 

t25$R2[pos] <- c(
  length(dat$Year), 
  rep(NA, 1),
  length(dat_amb$Year), 
  rep(NA, 1),
  length(dat_MID_ES$Year), 
  rep(NA, 1),
  length(dat_max_absES$Year), 
  rep(NA, 1),
  length(dat_max_meanES$Year), 
  rep(NA, 1),
  length(dat_max_ES$Year), 
  rep(NA, 1)) 

# adding model R2 estimates to the table
posR2 <- c(1, 4, 7, 10, 13, 16) #row positions
t25$`R2`[posR2] <- round(c(R2(MR_all_Year)[1], R2(MR_amb_Year)[1], R2(MR_MID_Year)[1], R2(MR_maxabsES_Year)[1], R2(MR_maxmeanES_Year)[1], R2(MR_maxES_Year)[1]),3)

write.csv(t25, file = "./Tables/Table_S25_MR_Year.csv", row.names = F, na = "")    
```


###################################################################    
#### Table S26      
Bayesian MA models results - all data and 5 data subsets    
Multivariate Multilevel Bayesian Phylogenetic Meta-Analysis (intercept-only) model. 

```{r Table MA Bayesian}
# loading data
load(file = "./Rdata/bayesian_MA1.1.Rdata")
load(file = "./Rdata/bayesian_MA1.2.Rdata")
load(file = "./Rdata/bayesian_MA1.3.Rdata")
m1 <- bayesian_MA1.1
m2 <- bayesian_MA1.2
m3 <- bayesian_MA1.3

#gelman.diag(list(m1$Sol[,1],m2$Sol[,1],m3$Sol[,1]))
#gelman.diag(list(m1$Deviance,m2$Deviance,m3$Deviance))
#check chain autocorrelation
#autocorr(m1$Sol[,c(3)])
#autocorr(m1$VCV[,c(1)])
#autocorr(m2$Sol[,c(3)])
#autocorr(m2$VCV[,c(1)])
#autocorr(m3$Sol[,c(3)])
#autocorr(m3$VCV[,c(1)])
#extract and summarise DIC
#print(c(m1$DIC,m2$DIC,m3$DIC))
#mean(c(m1$DIC, m2$DIC, m3$DIC)) 
#range(c(m1$DIC, m2$DIC, m3$DIC))  

chosen <- bayesian_MA1.1
summary(chosen)

### mean and range for DIC
DICs <- c(m1$DIC, m2$DIC, m3$DIC)
DICmean <- format(round(mean(DICs),2), nsmschool = 2)
DICrange <- paste0("(",format(round(range(DICs)[1],2), nsmschool = 2)," - ",format(round(range(DICs)[2],2), nsmschool = 2), ")")
DIC <- c("", "", "", DICmean, DICrange)
# heterogeneity and variances
WI <- na.omit(1/dat$Vd) 
s2I <- sum(WI*(length(WI)-1))/(sum(WI)^2-sum(WI^2)) # typical sampling (measurement) error variance (sigma2m)
total_var <- chosen$VCV[,"ArticleID"]+chosen$VCV[,"ExperimentID"]+chosen$VCV[,"units"]+s2I # total variance (sigma2t)
total_het <- (total_var-s2I)/total_var # total heterogeneity I2

ScalePoint <- round(c(posterior.mode(chosen$VCV[,"ScalePoint"]/total_var)*100, summary(chosen$VCV[,"ScalePoint"]/total_var)$statistics[1]*100, summary(chosen$VCV[,"ScalePoint"]/total_var)$statistics[2]*100),1)
ArticleID <- round(c(posterior.mode(chosen$VCV[,"ArticleID"]/total_var)*100, summary(chosen$VCV[,"ArticleID"]/total_var)$statistics[1]*100, summary(chosen$VCV[,"ArticleID"]/total_var)$statistics[2]*100),1)
ExperimentID <- round(c(posterior.mode(chosen$VCV[,"ExperimentID"]/total_var)*100, summary(chosen$VCV[,"ExperimentID"]/total_var)$statistics[1]*100, summary(chosen$VCV[,"ExperimentID"]/total_var)$statistics[2]*100),1)
units <- round(c(posterior.mode(chosen$VCV[,"units"]/total_var)*100, summary(chosen$VCV[,"units"]/total_var)$statistics[1]*100, summary(chosen$VCV[,"units"]/total_var)$statistics[2]*100),1)
total <- round(c(posterior.mode(total_het)*100, summary(total_het)$statistics[1]*100, summary(total_het)$statistics[2]*100),1)
var.res_df <- t(data.frame(DIC=DIC, ScalePoint=c(ScalePoint,"",""),ArticleID=c(ArticleID,"",""), ExperimentID=c(ExperimentID,"",""), units=c(units,"",""), total=c(total,"","")))

var.res_df <- format(var.res_df, nsmschool = 2)
colnames(var.res_df) <- c("Mode","Mean","SD", "DIC mean", "DIC range")
rownames(var.res_df) <- c("school data MR full model:","ScalePoint","I2Article","I2Experiment","I2residual","I2total")

#summary(chosen)

# fixed effects (intercept)
mod <- posterior.mode(chosen$Sol)[1]
av <- summary(chosen$Sol[,1:2])$statistics[,1][1]
sd <- summary(chosen$Sol[,1:2])$statistics[,2][1]
HPDlo <- HPDinterval(chosen$Sol[,1:2])[,1][1]
HPDhi <- HPDinterval(chosen$Sol[,1:2])[,2][1]
fixed.res_df <- data.frame(Mode=mod,Mean=av,SD=sd,Lower=HPDlo,Upper=HPDhi)
#fixed.res_df <- fixed.res_df[c(2,1,3,4,5),] #reorder moderators
fixed.res_df <- format(round(fixed.res_df,3), nsmuni = 3)
rownames(fixed.res_df) <- c("Intercept")

# random effects
Mode <- posterior.mode(chosen$VCV)
Mean <- summary(chosen$VCV)$statistics[,1]
SD <- summary(chosen$VCV)$statistics[,2]
Lower <- HPDinterval(chosen$VCV)[,1]
Upper <- HPDinterval(chosen$VCV)[,2]
random.res_df <- data.frame(Mode=Mode,Mean=Mean,SD=SD,Lower=Lower,Upper=Upper)
random.res_df <- random.res_df[-4,]
random.res_df <- format(round(random.res_df,3), nsmuni = 3)
rownames(random.res_df) <- c("ScalePoint","ArticleID","ExperimentID","residuals")

var.res_df
fixed.res_df
random.res_df

# loading table
t26 <- read_excel("./Tables/Results_Tables_v3.xlsx", sheet = 23)

# saving position of rows
pos <- c(2, 6:9, 13:17)
density.pos <- c(2, 6:9, 13)

# updating table
t26$`Fixed effects`[pos] <- c(fixed.res_df$Mode, random.res_df$Mode, as.numeric(var.res_df[2:6,1]))
t26$...3[pos] <- c(fixed.res_df$Mean, random.res_df$Mean, as.numeric(var.res_df[2:6,2]))
t26$...4[pos] <- c(fixed.res_df$SD, random.res_df$SD, as.numeric(var.res_df[2:6,3]))

t26$...5[density.pos] <- c(fixed.res_df$Lower, random.res_df$Lower, as.numeric(var.res_df[1,4]))
t26$...6[density.pos] <- c(fixed.res_df$Upper, random.res_df$Upper, var.res_df[1,5])

write.csv(t26, file = "./Tables/Table_S26_Bayesian_MA.csv", row.names = F, na = "")
```
